% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation_AM.R
\name{add.mean.generate.each}
\alias{add.mean.generate.each}
\title{Generate an additive mean function}
\usage{
add.mean.generate.each(j, Xi.each, Yspace, Ymu, Ydim = 1)
}
\arguments{
\item{Xi.each}{an \eqn{n} vector of scores supported on \eqn{[0,1]}.}

\item{Yspace}{the name of the underlying space \eqn{\mathcal{M}_Y} of \eqn{Y}.}

\item{Ymu}{an \eqn{m} vector of the Frechet mean of \eqn{Y}.}

\item{Ydim}{the intrinsic dimension of \eqn{\mathcal{M}_Y}.}
}
\value{
an \eqn{n\times m} matrix of an additive mean \eqn{\mathfrak{m}: [0,1] \to T_{\mu_Y}\mathcal{M}_Y}.
}
\description{
Generate an additive mean function \eqn{\mathfrak{m}: [0,1] \to T_{\mu_Y}\mathcal{M}_Y}, where \eqn{\mu_Y} is the Frechet mean of \eqn{Y}.
}
