% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_BayesHilbert.R
\name{basis.BayesHilbert}
\alias{basis.BayesHilbert}
\title{Basis on the tangent space at the point for Bayes-Hilbert data}
\usage{
basis.BayesHilbert(p, dim = 50)
}
\description{
Basis on the tangent space at the point for Bayes-Hilbert data
}
