% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_Euclid.R
\name{basis.Euclid}
\alias{basis.Euclid}
\title{Basis on the tangent space at the point for Euclid data}
\usage{
basis.Euclid(p, dim = NULL)
}
\description{
Basis on the tangent space at the point for Euclid data
}
