% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_functional.R
\name{basis.functional}
\alias{basis.functional}
\title{Basis on the tangent space at the point for functional data}
\usage{
basis.functional(p, dim = 50)
}
\description{
Basis on the tangent space at the point for functional data
}
