% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_AffInv.R, R/Geometry_SPD_LogEuclid.R,
%   R/Geometry_Wasserstein.R, R/Geometry_functional.R, R/Geometry_simplex.R,
%   R/Geometry_sphere.R
\name{basis.manifold}
\alias{basis.manifold}
\alias{basis.BayesHilbert}
\alias{basis.Euclid}
\alias{basis.SPD.AffInv}
\alias{basis.SPD.LogEuclid}
\alias{basis.Wasserstein}
\alias{basis.functional}
\alias{basis.simplex}
\alias{basis.sphere}
\title{Orthonormal basis on the tangent space of the manifold}
\usage{
basis.manifold(p, dim = 50, space = "Euclid")

basis.BayesHilbert(p, dim = 50)

basis.Euclid(p, dim = NULL)

basis.SPD.AffInv(p, dim = 50)

basis.SPD.LogEuclid(p, dim = 50)

basis.Wasserstein(p, dim = 50)

basis.functional(p, dim = 50)

basis.simplex(p, dim = NULL)

basis.sphere(p, dim = NULL)
}
\arguments{
\item{p}{an \eqn{m} vector representing the base point in \eqn{\mathcal{M}}.}

\item{dim}{the number of orthonormal basis, only used for infinite-dimensional \eqn{\mathcal{M}}.}

\item{space}{the name of space, which must be one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n\times m} matrix where each row is an orthonormal basis of the tangent space \eqn{T_p\mathcal{M}}.
}
\description{
Compute the orthonormal basis of the tangent space \eqn{T_p\mathcal{M}}.
}
\section{Functions}{
\itemize{
\item \code{basis.BayesHilbert()}: Method

\item \code{basis.Euclid()}: Method

\item \code{basis.SPD.AffInv()}: Method

\item \code{basis.SPD.LogEuclid()}: Method

\item \code{basis.Wasserstein()}: Method

\item \code{basis.functional()}: Method

\item \code{basis.simplex()}: Method

\item \code{basis.sphere()}: Method

}}
