% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_simplex.R
\name{basis.simplex}
\alias{basis.simplex}
\title{Basis on the tangent space at the point for simplex data}
\usage{
basis.simplex(p, dim = NULL)
}
\description{
Basis on the tangent space at the point for simplex data
}
