% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_sphere.R
\name{basis.sphere}
\alias{basis.sphere}
\title{Basis on the tangent space at the point for sphere data}
\usage{
basis.sphere(p, dim = NULL)
}
\description{
Basis on the tangent space at the point for sphere data
}
