% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_Wasserstein.R
\name{basis.Wasserstein}
\alias{basis.Wasserstein}
\title{Basis on the tangent space at the point for Wasserstein data}
\usage{
basis.Wasserstein(p, dim = 50)
}
\description{
Basis on the tangent space at the point for Wasserstein data
}
