% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{Check.manifold}
\alias{Check.manifold}
\title{Check if the given space is a manifold}
\usage{
Check.manifold(space)
}
\arguments{
\item{space}{the name of space, which must be one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\description{
Check if the given space is a manifold.
The space name must be one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.
}
\details{
Basic functions for manifolds

Basic geometry structure:
inner product, norm in tangent spaces, distance
exponential and logarithmic maps

For PCA:
Frechet mean, PCA
}
