% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_simplex.R
\name{clr}
\alias{clr}
\title{Centered log-ratio transform}
\usage{
clr(x)
}
\arguments{
\item{x}{an (n,m) matrix of simplex-valued or Bayes-Hilbert space-valued data}
}
\value{
A clr transformed (n,m) matrix
}
\description{
Transform simplex or density data to R^m
clr(x)_i = log(x_i)-sum_j log(x_j)
}
