% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation.R
\name{covariates.generate.each}
\alias{covariates.generate.each}
\title{Generate a manifold-valued covariate}
\usage{
covariates.generate.each(Xi, dim, space = "Euclid")
}
\arguments{
\item{Xi}{an \eqn{n\times m'} score matrix, where \eqn{m'} is the intrinsic dimension (the number of scores).}

\item{dim}{the dimension of \eqn{X}.}

\item{space}{the underlying space, see \code{\link{Check.manifold}}.}
}
\value{
an \eqn{n\times m} matrix. Each row represents a manifold-valued covariate.
}
\description{
Given \eqn{j}, generate manifold-valued covariates \eqn{X_{ij}} for \eqn{1\le i\le n}.
}
