% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_AffInv.R, R/Geometry_SPD_LogEuclid.R,
%   R/Geometry_Wasserstein.R, R/Geometry_functional.R, R/Geometry_simplex.R,
%   R/Geometry_sphere.R
\name{dist.manifold}
\alias{dist.manifold}
\alias{dist.BayesHilbert}
\alias{dist.Euclid}
\alias{dist.SPD.AffInv}
\alias{dist.SPD.LogEuclid}
\alias{dist.Wasserstein}
\alias{dist.functional}
\alias{dist.simplex}
\alias{dist.sphere}
\title{Geodesic distance on the manifold}
\usage{
dist.manifold(p, q, space = "Euclid")

dist.BayesHilbert(p, q)

dist.Euclid(p, q)

dist.SPD.AffInv(p, q)

dist.SPD.LogEuclid(p, q)

dist.Wasserstein(p, q)

dist.functional(p, q)

dist.simplex(p, q)

dist.sphere(p, q)
}
\arguments{
\item{p, q}{\eqn{m} vectors or \eqn{n\times m} matrices, where each row is a point on \eqn{\mathcal{M}}.}

\item{space}{the name of space, which must be one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n} vector of geodesic distances between corresponding rows of \eqn{p} and \eqn{q}.
}
\description{
Compute the geodesic distance on the manifold \eqn{\mathcal{M}}.
}
\section{Functions}{
\itemize{
\item \code{dist.BayesHilbert()}: Method

\item \code{dist.Euclid()}: Method

\item \code{dist.SPD.AffInv()}: Method

\item \code{dist.SPD.LogEuclid()}: Method

\item \code{dist.Wasserstein()}: Method

\item \code{dist.functional()}: Method

\item \code{dist.simplex()}: Method

\item \code{dist.sphere()}: Method

}}
