% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensorProduct.R
\name{dist.tensor}
\alias{dist.tensor}
\title{Distance on the Tensor Product Space}
\usage{
dist.tensor(x, y)
}
\arguments{
\item{x, y}{elements in \eqn{\mathbb{H}_1\otimes \mathbb{H}_2}, created by \code{\link{make.tensor}}.}
}
\value{
the distance between x and y.
}
\description{
Computes the distance between two elements in \eqn{\mathbb{H}_1\otimes \mathbb{H}_2}.
}
