% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_data_generation.R
\name{error.generate}
\alias{error.generate}
\title{Gerenate random errors on the tangent space at mu.}
\usage{
error.generate(n, space, dim, error.rho = 0.5, error.std = 1)
}
\arguments{
\item{n}{a number of data.}

\item{space}{an underlying space.}

\item{dim}{an intrinsic dimension of the tangent space.}

\item{error.rho}{a correlation parameter.}

\item{error.std}{a standard deviation parameter.}
}
\value{
an \eqn{n\times m} matrix of generated random error.
}
\description{
Gerenate random errors on the tangent space at mu.
}
