% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_BayesHilbert.R
\name{FrechetMean.BayesHilbert}
\alias{FrechetMean.BayesHilbert}
\title{Frechet mean for Bayes-Hilbert data}
\usage{
FrechetMean.BayesHilbert(X)
}
\description{
Frechet mean for Bayes-Hilbert data
}
