% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{FrechetMean.manifold}
\alias{FrechetMean.manifold}
\title{Frechet mean on the manifold}
\usage{
FrechetMean.manifold(X, space = "Euclid")
}
\arguments{
\item{X}{a matrix with dimension (n,m). Each row is a point on the Riemannian metric spaces.}

\item{space}{a name of space, one of 'Euclid', 'simplex', 'sphere', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an m vector
}
\description{
Frechet mean on the manifold
}
