% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_oracle_CV.R
\name{get.loss.LR.oracle}
\alias{get.loss.LR.oracle}
\title{Compute loss for each Xdim.max}
\usage{
get.loss.LR.oracle(X, LogY, Xnew, LogYnew, Ymu, Xdim.max, proper.indices)
}
\arguments{
\item{X}{a \eqn{p} list of score matrix. Each score matrix is an \eqn{n}-by-\eqn{d_j} matrix, see \code{\link{PCA.manifold.list}}.}

\item{LogY}{an \eqn{n}-by-\eqn{m} response matrix.}

\item{Xnew}{a new \eqn{p} list of score matrix. Each score matrix is an \eqn{n'}-by-\eqn{d_j} matrix.}

\item{LogYnew}{a new \eqn{n}-by-\eqn{m} response matrix.}

\item{Ymu}{an underlying point of the tangent space of LogY and LogYnew.}

\item{Xdim.max}{a max dimension of \eqn{X_j}, real>0.}

\item{proper.indices}{a \eqn{s} vector of indices of relevant \eqn{X_j}.}
}
\value{
a prediction loss.
}
\description{
Compute loss for each Xdim.max
}
