% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBF_comp.R
\name{get.min.bandwidths.kfold}
\alias{get.min.bandwidths.kfold}
\title{Minimal Bandwidths for K-fold CV}
\usage{
get.min.bandwidths.kfold(X.list, degree = 0, const = 1.001)
}
\arguments{
\item{X.list}{a list of \eqn{n\times p} score matrices supported \eqn{[0,1]}.}

\item{degree}{the degree of the local polynomial fitting. Options are: 0 for Nadaraya-Watson (default), 1 for local-linear, etc.}

\item{const}{a constant to adjust the minimal bandwidths.}
}
\value{
a \eqn{p} vector of adjusted minimal bandwidths.
}
\description{
Computes the minimal bandwidths to ensure that one-dimensional KDEs are invertible on the interval \eqn{[0,1]}.
The function returns a vector of minimal bandwidths, each multiplied by an adjustment constant.
}
