% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_GCV.R
\name{GLM.GCV}
\alias{GLM.GCV}
\title{Generalized cross validation for high-dimensional linear regression}
\usage{
GLM.GCV(
  Xorg,
  Yorg,
  Xorgnew,
  Yorgnew,
  lambda.list,
  Xdim.max.list,
  R.list,
  penalty = "LASSO",
  link = "binomial",
  phi = 1,
  gamma = 0,
  max.cv.iter = 20,
  cv.threshold = 1e-10,
  eta = 0.001,
  max.iter = 500,
  threshold = 1e-10
)
}
\arguments{
\item{Xorg}{a list of manifold-valued covariates, see \code{\link{PCA.manifold.list}}.}

\item{Yorg}{an \eqn{n\times m} response matrix.}

\item{lambda.list}{a vector of lambda.}

\item{Xdim.max.list}{a vector of max dimension of \eqn{X_j}.}

\item{R.list}{a vector of constrained bound.}

\item{penalty}{a method of penalty. It should be one of 'LASSO', 'SCAD', or 'MCP'.}

\item{phi}{a parameter in computing ADMM-MM algorithm for the majorized objective function, default 1.}

\item{gamma}{a parameter for SCAD (3.7) or MCP (3), parentheses: default value.}

\item{max.cv.iter}{a number of maximum CV iterations, default 20.}

\item{cv.threshold}{a parameter to modify the computation error in CV, default 1e-10.}

\item{eta}{a parameter in computing ADMM-MM algorithm for the proximal norm square, default 1e-3.}

\item{max.iter}{a maximum iteration, default 500.}

\item{threshold}{an algorihtm convergence threshold, default 1e-10.}

\item{Xnew}{a list of new covariate observations.}

\item{Ynew}{an \eqn{n'\times} new response observations.}

\item{Yspace}{an underlying space of \eqn{Y}.}
}
\value{
an \code{\link{LM}} object.
   \describe{
      \item{parameter.list}{a list of optimal parameters for each CV step.}
      \item{loss.list}{a list of loss for each CV step.}
      \item{runtime}{running time}
      \item{...}{see \code{\link{LM}}.}
}
}
\description{
GCV for a LM object
It is based on CBS algorithm and uses a function LM_GCV in LM_GCV.cpp.
}
