% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_oracle_GCV.R
\name{GLM.oracle.kfold}
\alias{GLM.oracle.kfold}
\title{Kfold cross validation for a GLM.oracle function}
\usage{
GLM.oracle.kfold(
  Xall,
  Yall,
  kfold,
  Xdim.max.list,
  proper.indices = NULL,
  link = "binomial",
  seed = NULL,
  max.cv.iter = 20,
  cv.threshold = 1e-10,
  eta = 0.001,
  max.iter = 500,
  threshold = 1e-10
)
}
\arguments{
\item{Xall}{a list of manifold-valued covariates, see \code{\link{PCA.manifold.list}}.}

\item{Yall}{an \eqn{n\times m} response matrix.}

\item{kfold}{a number of kfold CV, int>0.}

\item{Xdim.max.list}{a vector of max dimension of \eqn{X_j}.}

\item{proper.indices}{a vector of indices of relevant \eqn{X_j}.}

\item{seed}{a random seed, int>0, default: non-random (NULL).}

\item{Yspace}{an underlying space of Yorg and Ynew.}
}
\value{
an \eqn{\code{LM}} object
   \describe{
      \item{Xdim.max.list}{a vector of Xdim.max.}
      \item{loss.list}{a list of loss for each CV step.}
      \item{runtime}{running time.}
      \item{...}{see \code{\link{LM.oracle}}.}
}
}
\description{
Kfold cross validation for a GLM.oracle function
}
