% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_BayesHilbert.R
\name{inner.BayesHilbert}
\alias{inner.BayesHilbert}
\title{Inner product on tangent space at p for Bayes-Hilbert data}
\usage{
inner.BayesHilbert(u, v, p = NULL)
}
\description{
Inner product on tangent space at p for Bayes-Hilbert data
}
