% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_Euclid.R
\name{inner.Euclid}
\alias{inner.Euclid}
\title{Inner product on tangent space at p for Euclid data}
\usage{
inner.Euclid(u, v, p = NULL)
}
\description{
Inner product on tangent space at p for Euclid data
}
