% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_functional.R
\name{inner.functional}
\alias{inner.functional}
\title{Inner product on tangent space at p for functional data}
\usage{
inner.functional(u, v, p = NULL)
}
\description{
Inner product on tangent space at p for functional data
}
