% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_AffInv.R, R/Geometry_SPD_LogEuclid.R,
%   R/Geometry_Wasserstein.R, R/Geometry_functional.R, R/Geometry_simplex.R,
%   R/Geometry_sphere.R
\name{inner.manifold}
\alias{inner.manifold}
\alias{inner.BayesHilbert}
\alias{inner.Euclid}
\alias{inner.SPD.AffInv}
\alias{inner.SPD.LogEuclid}
\alias{inner.Wasserstein}
\alias{inner.functional}
\alias{inner.simplex}
\alias{inner.sphere}
\title{Inner product on the tangent space}
\usage{
inner.manifold(u, v, p, space = "Euclid")

inner.BayesHilbert(u, v, p = NULL)

inner.Euclid(u, v, p = NULL)

inner.SPD.AffInv(u, v, p = NULL)

inner.SPD.LogEuclid(u, v, p = NULL)

inner.Wasserstein(u, v, p = NULL)

inner.functional(u, v, p = NULL)

inner.simplex(u, v, p = NULL)

inner.sphere(u, v, p = NULL)
}
\arguments{
\item{u, v}{\eqn{n\times m'} matrices, where each row is a tangent vector on \eqn{T_p\mathcal{M}}.}

\item{p}{an \eqn{m} vector representing the base point in \eqn{\mathcal{M}}.}

\item{space}{the name of space, which must be one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n} vector of inner products for each corresponding row of \eqn{u} and \eqn{v}.
}
\description{
Compute the inner product on the tangent space \eqn{T_p\mathcal{M}} of the manifold \eqn{\mathcal{M}}.
}
\section{Functions}{
\itemize{
\item \code{inner.BayesHilbert()}: Method

\item \code{inner.Euclid()}: Method

\item \code{inner.SPD.AffInv()}: Method

\item \code{inner.SPD.LogEuclid()}: Method

\item \code{inner.Wasserstein()}: Method

\item \code{inner.functional()}: Method

\item \code{inner.simplex()}: Method

\item \code{inner.sphere()}: Method

}}
