% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_simplex.R
\name{inner.simplex}
\alias{inner.simplex}
\title{Inner product on tangent space at p for simplex data}
\usage{
inner.simplex(u, v, p = NULL)
}
\description{
Inner product on tangent space at p for simplex data
}
