% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_sphere.R
\name{inner.sphere}
\alias{inner.sphere}
\title{Inner product on tangent space at p for sphere data}
\usage{
inner.sphere(u, v, p = NULL)
}
\description{
Inner product on tangent space at p for sphere data
}
