% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_Wasserstein.R
\name{inner.Wasserstein}
\alias{inner.Wasserstein}
\title{Inner product on tangent space at p for Wasserstein data}
\usage{
inner.Wasserstein(u, v, p = NULL)
}
\description{
Inner product on tangent space at p for Wasserstein data
}
