% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_BayesHilbert.R
\name{inv.clr.BayesHilbert}
\alias{inv.clr.BayesHilbert}
\title{Inverse centered log-transform for density data}
\usage{
inv.clr.BayesHilbert(x, normalizer = 1)
}
\arguments{
\item{x}{an \eqn{n\times m} matrix of the clr-transformed data.}

\item{normalizer}{a normalization constant, with a default value of 1.}
}
\value{
a \eqn{n\times m} matrix of density data.
}
\description{
Compute the inverse centered log-ratio transform to convert data from \eqn{\mathbb{R}^m} to density data, see \code{\link{clr}}.
Specifically, for each \eqn{x=(x_i)\in\mathbb{R}^m}, the transformation is give by
\deqn{\text{inv.clr}(x)_i = \frac{\exp(x_i)}{\frac{1}{m}\sum_{j} \exp(x_j)} \times \text{normalizer}}.
}
