% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_simplex.R
\name{inv.clr.simplex}
\alias{inv.clr.simplex}
\title{Inverse centered log-transform for simplex data}
\usage{
inv.clr.simplex(x, normalizer = 1)
}
\arguments{
\item{x}{A clr transformed (n,m) matrix}

\item{normalizer}{A normalizing constant, default=1}
}
\value{
A simplex-valued (n,m) matrix
}
\description{
Inverse transform R^m to simplex data
inv.clr(x)_i = exp(x_i)/(sum_j exp(x_i)) * normalizer
}
