% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBF_comp.R
\name{KDE}
\alias{KDE}
\title{Kernel Density Estimator}
\usage{
KDE(
  X,
  bandwidths,
  degree = 0,
  ngrid = 51,
  Kdenom_method = "numeric",
  is.proj = FALSE
)
}
\arguments{
\item{X}{an \eqn{n\times p} score matrix.}

\item{bandwidths}{a \eqn{p} vector of bandwidths.}

\item{degree}{the degree of the local polynomial fitting. Options are: 0 for Nadaraya-Watson (default), 1 for local-linear, etc.}

\item{ngrid}{the number of grid points for evaluating the KDE (default: 51).}

\item{Kdenom_method}{a method used for the denominator in the normalized kernel calculation.}

\item{is.proj}{logical. If \code{TRUE}, this function gives the projection operators.}
}
\value{
a \code{KDE} object with the following components: 
\describe{
      \item{kde.1d}{a \eqn{p} list of one-dimensional KDE where each element is (ngrid,degree+1,degree+1) array.}
      \item{kde.2d}{a \eqn{(p*p*ngrid*ngrid,degree+1,degree+1)} array of two-dimensional KDE. This only exists when \code{is.proj=FALSE}.}
      \item{proj}{a \eqn{(p\times p\times ngrid\times ngrid) \times (degree+1) \times (degree+1)}-dimensional array. This only exists when \code{is.proj=TRUE}.}
      \item{kde.1d.inv}{a \eqn{p} list of inverse matrices of one-dimensional KDE where each element is (ngrid,degree+1,degree+1) array.}
      \item{grids}{grid points.}
      \item{weights}{a vector of weights for fast computing numerical integration.}
}
}
\description{
Computes one- and two-dimensional kernel density estimators for \eqn{X_j} on the interval \eqn{[0,1]} for \eqn{1\le j\le p}.
This function also computes the projection operator \eqn{\Pi_{jk}:L^2(f_k) \to L_2(f_j)} for \eqn{1\le j\neq k\le p}.
It uses the normalized Epanechnikov kernel for computing the KDEs.
}
