% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_data_generation.R
\name{linear.data.generate}
\alias{linear.data.generate}
\title{Generate linear regression data}
\usage{
linear.data.generate(
  n,
  Xspaces,
  Yspace,
  Xdims,
  Ydim,
  proper.indices = NULL,
  beta.norm = 1,
  Xrho = 0.5,
  Xsigma = 1,
  error.rho = 0.5,
  error.std = 1,
  seed = 1
)
}
\arguments{
\item{n}{a number of observation.}

\item{Xspaces}{a \eqn{p} vector of underlying spaces of \eqn{X_j}.}

\item{Yspace}{an underlying space of \eqn{Y}.}

\item{Xdims}{a \eqn{p} vector of intrinsic dimensions of \eqn{H_j}.}

\item{Ydim}{an intrinsic dimension of \eqn{H_Y}.}

\item{proper.indices}{an indices of nonzero operators.}

\item{beta.norm}{the tensor norm of nonzero operators.}

\item{Xrho}{a correlation parameter for \eqn{X_j}.}

\item{Xsigma}{a variance parameter for \eqn{X_j}.}

\item{error.rho}{a correlation parameter for error.}

\item{error.std}{a standard deviation parameter for error.}

\item{seed}{a random seed (1<=seed)}
}
\value{
a list of generated data
   \describe{
      \item{X}{a \eqn{p} list of generated covariates. Each \eqn{X_j} is an \eqn{n\times Xdim_j} matrix.}
      \item{Y}{an \eqn{n\times m} matrix of response.}
      \item{beta}{a \eqn{p} list of generated operators.}
      \item{error}{a generated random error.}
      \item{Xbeta.each}{a \eqn{p} list of \eqn{B_j(X_j)}.}
      \item{Xbeta}{a sum of Xbeta.each.}
      \item{Ymu}{the Frechet mean of \eqn{Y}.}
      \item{...}{the other input parameters.}
}
}
\description{
Generate linear regression data.
The seed for generating beta is fixed as zero, so each beta is equal in simulations.
}
