% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation.R
\name{LM.data.generate}
\alias{LM.data.generate}
\title{Generate linear regression data}
\usage{
LM.data.generate(
  n,
  Xspaces,
  Yspace,
  Xdims,
  Ydim,
  proper.indices = NULL,
  beta.norm = 1,
  Xrho = 0.5,
  Xsigma = 1,
  error.rho = 0.5,
  error.std = 1,
  ngrid = 100,
  seed = 1
)
}
\arguments{
\item{n}{the number of data points}

\item{Xspaces}{a \eqn{p} vector of the names of the underlying spaces \eqn{\mathcal{M}_j} of \eqn{X_j}.}

\item{Yspace}{the name of the underlying space \eqn{\mathcal{M}_Y} of \eqn{Y}.}

\item{Xdims}{a \eqn{p} vector of the intrinsic dimension of \eqn{\mathcal{M}_j}.}

\item{Ydim}{the intrinsic dimension of \eqn{\mathcal{M}_Y}.}

\item{proper.indices}{an index set \eqn{\mathcal{S}=\{1\le j\le p : \mathfrak{B}_j\neq0\}}.}

\item{beta.norm}{a \eqn{p} vector (or a single integer) of the tensor norms, used only for \eqn{j\notin\mathcal{S}}. If an integer is provided, the same value is applied for all \eqn{X_j}.}

\item{Xrho}{a correlation parameter for \eqn{X_j} ranging from -1 to 1, with a default value of 0.5.}

\item{Xsigma}{a common standard deviation parameter for \eqn{X_j}, with a default value of 1.}

\item{error.rho}{a correlation parameter for \eqn{\varepsilon}.}

\item{error.std}{a standard deviation parameter for \eqn{\varepsilon}.}

\item{ngrid}{the number of grids. This is only used for infinite-dimensional spaces, see \code{\link{Check.manifold}}: "functional", "BayesHilbert" and "Wasserstein" spaces.}

\item{seed}{a random seed, which must be greater than 1.}
}
\value{
a list of data containing:
   \describe{
      \item{X}{a list of manifold-valued covariates, see \code{\link{covariates.generate}}.}
      \item{Y}{an \eqn{n\times m} matrix of manifold-valued responses.}
      \item{p}{the number of \eqn{X_j}.}
      \item{beta}{a \eqn{p} list of Hilbert-Schmidt operators, see \code{\link{tensor.beta.generate}}.}
      \item{Ymu}{the Frechet mean \eqn{\mu_Y} of \eqn{Y}.}
      \item{Xmu}{a \eqn{p} list of the Frechet means \eqn{\mu_j} of \eqn{X_j}.}
      \item{LogY}{the Riemannian logarithmic transformation \eqn{Log_{\mu_Y}Y}.}
      \item{LogX}{a \eqn{p} list of \eqn{Log_{\mu_j}X_j}, the Riemannian logarithmic transformations of \eqn{X_j}.}
      \item{Xbeta.each}{a \eqn{p} list of \eqn{\mathfrak{B}_j(Log_{\mu_j}X_j)}.}
      \item{Xbeta}{the sum of Xbeta.each, i.e., \eqn{\sum_{j=1}^p\mathfrak{B}_j(Log_{\mu_j}X_j)}.}
      \item{ExpXbeta}{the Riemannian exponential of Xbeta, i.e., \eqn{Exp_{\mu_Y}\Big(\sum_{j=1}^p\mathfrak{B}_j(Log_{\mu_j}X_j)\Big)}.}
      \item{...}{other input parameters.}
}
}
\description{
Generate linear regression data.
The seed for generating Hilbert-Schmidt operators \eqn{\mathfrak{B}_j} is fixed as zero, ensuring that each \eqn{\mathfrak{B}_j} is the same across simulations.
}
