% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_oracle_CV.R
\name{LM.oracle.kfold}
\alias{LM.oracle.kfold}
\title{Kfold Cross-Validation for Oracle Hilbert-Schmidt Linear Models}
\usage{
LM.oracle.kfold(
  Xorg,
  Yorg,
  Yspace,
  kfold,
  Xdim.max.list,
  proper.indices = NULL,
  seed = NULL
)
}
\arguments{
\item{Xorg}{a list of covariates with the following components (see also \code{\link{covariates.generate}}):
\describe{
      \item{j}{a \eqn{p} list of manifold-valued covariates, where each \eqn{j}th element is an \eqn{n\times T_j} matrix.}
      \item{spaces}{a \eqn{p} vector of the underlying spaces \eqn{\mathcal{M}_j} of \eqn{X_j}.}
      \item{p}{the number of \eqn{X_j}.}
}}

\item{Yorg}{an \eqn{n\times m} matrix of manifold-valued responses.}

\item{Yspace}{the name of the underlying space \eqn{\mathcal{M}_Y} of \eqn{Y}.}

\item{kfold}{the number of kfolds (default: 5).}

\item{Xdim.max.list}{a vector of the maximum dimension to which \eqn{X_j} will be reduced.}

\item{proper.indices}{an index set \eqn{\mathcal{S}=\{1\le j\le p : \mathfrak{B}_j\neq0\}}.}

\item{seed}{a random seed for kfolds. When it is not provided, the data will be partitioned in index order.}
}
\value{
an \code{\link{LM}} object with the following components:
   \describe{
      \item{pca}{a 'PCA.manifold.list' object, see \code{\link{PCA.manifold.list}}.}
      \item{Ymu}{the Frechet mean \eqn{\mu_Y} of \eqn{Y}.}
      \item{beta}{a \eqn{L_+^{*} \times m} matrix of estimated \eqn{\bm{\beta}}, where \eqn{L_+^{*}=\sum_{j=1}^p L_j^*} and \eqn{m} is the intrinsic dimension of \eqn{T_{\mu_Y}\mathcal{M}_Y}.}
      \item{beta.each}{a \eqn{p} list of \eqn{L_j^*\times m} matrices of \eqn{\bm{\beta}_j}.}
      \item{beta.norm}{a \eqn{p} vector of norms of \eqn{\bm{\beta}_j}.}
      \item{beta.vectors}{a \eqn{p} list of orthonormal bases of \eqn{X_j} obtained by \code{\link{PCA.manifold.list}}. Each basis is an \eqn{L_j^*\times T_j} matrix.}
      \item{beta.tensor}{a \eqn{p} list of estimated Hilbert-Schmidt operators, see \code{\link{make.tensor}}.}
      \item{proper.indices}{an index set an index set \eqn{\mathcal{S}=\{1\le j\le p : {\mathfrak{B}}_j\neq0\}}.}
      \item{parameter.list}{a list of optimal parameters for each CV update.}
      \item{loss.list}{a list of loss for each CV update.}
      \item{runtime}{the running time.}
      \item{...}{other parameters.}
}
}
\description{
Implements a Kfold cross-validation (Kfold-CV) for oracle Hilbert-Schmidt linear models.
For a more detailed description of parameters, see \code{\link{LM.oracle}}.
}
