% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_kfold.R
\name{LR.kfold}
\alias{LR.kfold}
\title{Kfold cross validation for high-dimensional linear regression for manifold-valued data}
\usage{
LR.kfold(
  Xall,
  Yall,
  Yspace,
  kfold,
  lambda.list,
  Xdim.max.list,
  R.list,
  phi = 1,
  penalty,
  gamma = 0,
  seed = NULL,
  max.cv.iter = 20,
  cv.threshold = 1e-10,
  eta = 0.001,
  max.iter = 500,
  threshold = 1e-10
)
}
\arguments{
\item{Xall}{a list of manifold-valued covariates, see \code{\link{PCA.manifold.list}}.}

\item{Yall}{an \eqn{n\times m} response matrix.}

\item{Yspace}{an underlying space of \eqn{Y}.}

\item{kfold}{a number of kfold CV, int>0.}

\item{lambda.list}{a vector of lambda.}

\item{Xdim.max.list}{a vector of max dimension of \eqn{X_j}.}

\item{R.list}{a vector of constrained bound.}

\item{phi}{a parameter in computing ADMM-MM algorithm for the majorized objective function, default 1.}

\item{penalty}{a method of penalty. It should be one of 'LASSO', 'SCAD', or 'MCP'.}

\item{gamma}{a parameter for SCAD (3.7) or MCP (3), parentheses: default value.}

\item{seed}{a random seed, int>0, default: non-random (NULL).}

\item{max.cv.iter}{a number of maximum CV iterations, default 20.}

\item{cv.threshold}{a parameter to modify the computation error in CV, default 1e-10.}

\item{eta}{a parameter in computing ADMM-MM algorithm for the proximal norm square, default 1e-3.}

\item{max.iter}{a maximum iteration, default 500.}

\item{threshold}{an algorihtm convergence threshold, default 1e-10.}
}
\value{
an \code{\link{LR}} object.
   \describe{
      \item{parameter.list}{a list of optimal parameters for each CV step.}
      \item{loss.list}{a list of loss for each CV step.}
      \item{runtime}{running time}
      \item{...}{see \code{\link{LR}}.}
}
}
\description{
Kfold CV for an \code{\link{LR}} function.
It is based on CBS algorithm and uses a function LR_kfold in LR_kfold.cpp.
}
