% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_oracle_CV.R
\name{LR.oracle.GCV}
\alias{LR.oracle.GCV}
\title{GCV for an LR.oracle function}
\usage{
LR.oracle.GCV(
  Xorg,
  Yorg,
  Xorgnew,
  Yorgnew,
  Yspace,
  Xdim.max.list,
  proper.indices
)
}
\arguments{
\item{Xorg}{a list of manifold-valued covariates, see \code{\link{PCA.manifold.list}}.}

\item{Yorg}{an \eqn{n\times m} response matrix.}

\item{Yspace}{an underlying space of Yorg and Ynew.}

\item{Xdim.max.list}{a vector of max dimension of \eqn{X_j}.}

\item{proper.indices}{a vector of indices of relevant \eqn{X_j}.}

\item{Xnew}{a new list of manifold-valued covariates.}

\item{Ynew}{a new \eqn{n'\times m} response matrix.}
}
\value{
an \eqn{\code{LR}} object
   \describe{
      \item{Xdim.max.list}{a vector of Xdim.max.}
      \item{loss.list}{a list of loss for each CV step.}
      \item{runtime}{running time.}
      \item{...}{see \code{\link{LR.oracle}}.}
}
}
\description{
GCV for an LR.oracle function
}
