% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_oracle_CV.R
\name{LR.oracle.kfold}
\alias{LR.oracle.kfold}
\title{Kfold cross validation for an LR.oracle function}
\usage{
LR.oracle.kfold(
  Xall,
  Yall,
  Yspace,
  kfold,
  Xdim.max.list,
  proper.indices,
  seed = NULL
)
}
\arguments{
\item{Xall}{a \eqn{p} list of manifold-valued covariates. Each \eqn{X_j} is an \eqn{n}-by-\eqn{d_j} matrix, see \code{\link{PCA.manifold.list}}.}

\item{Yall}{an \eqn{n}-by-\eqn{m} response matrix.}

\item{Yspace}{an underlying space of Yorg and Ynew.}

\item{kfold}{a number of kfold CV, int>0.}

\item{Xdim.max.list}{a vector of max dimension of \eqn{X_j}.}

\item{proper.indices}{a \eqn{s} vector of indices of relevant \eqn{X_j}.}

\item{seed}{a random seed, int>0, default: non-random (NULL).}
}
\value{
an \eqn{\code{LR}} object
   \describe{
      \item{Xdim.max.list}{a vector of Xdim.max.}
      \item{loss.list}{a list of loss for each CV step.}
      \item{runtime}{running time.}
      \item{...}{see \code{\link{LR.oracle}}.}
}
}
\description{
Kfold cross validation for an LR.oracle function
}
