% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_data_generation.R
\name{make.nonsym.Idmat}
\alias{make.nonsym.Idmat}
\title{Generate a non-symmetric identity matrix.
given p1,p2, the matrix I is a p1*p2 matrix with I_{jk}=delta_{jk}}
\usage{
make.nonsym.Idmat(p1, p2)
}
\description{
Generate a non-symmetric identity matrix.
given p1,p2, the matrix I is a p1*p2 matrix with I_{jk}=delta_{jk}
}
