% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensorProduct.R
\name{make.tensor}
\alias{make.tensor}
\title{Make an element in tensor product spaces of \eqn{H_1} and \eqn{H_2}, \eqn{H_1\otimes H_2}}
\usage{
make.tensor(element1, element2, space1, space2, mu1, mu2)
}
\arguments{
\item{element1}{an \eqn{n}-by-\eqn{m_1} matrix. Each row is an element in \eqn{H_1}.}

\item{element2}{an \eqn{n}-by-\eqn{m_2} matrix. Each row is an element in \eqn{H_2}.}

\item{space1, space2}{the underlying spaces of \eqn{H_1} and \eqn{H_2}, respectively.}

\item{mu1, mu2}{base elements in \eqn{H_1} and \eqn{H_2}, respectively. These will be used for computing norms.}
}
\value{
a tensor product object, class "tensor"
   \describe{
      \item{n}{a number of elements \eqn{n}.}
      \item{mu1,mu2}{base elements in \eqn{H_1} and \eqn{H_2}, respectively.}
      \item{space1,space2}{the underlying space of \eqn{H_1} and \eqn{H_2}, respectively.}
      \item{element1}{an \eqn{n}-by-\eqn{m_1} matrix. Each row is an element in \eqn{H_1}.}
      \item{element2}{an \eqn{n}-by-\eqn{m_2} matrix. Each row is an element in \eqn{H_2}.}
      \item{dim}{the intrinsic input dimension of \eqn{H_1}.}
}
}
\description{
Make an element in tensor product spaces of \eqn{H_1} and \eqn{H_2}, \eqn{H_1\otimes H_2}
Let \eqn{d_1} and \eqn{d_2} be the dimensions of \eqn{H_1} and \eqn{H_2}.
}
