\name{NonEuclidReg-package}
\alias{NonEuclidReg-package}
\alias{NonEuclidReg}
\docType{package}
\title{
\packageTitle{NonEuclidReg}
}
\description{
\packageDescription{NonEuclidReg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NonEuclidReg}
\packageIndices{NonEuclidReg}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NonEuclidReg}

Maintainer: \packageMaintainer{NonEuclidReg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
