% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_BayesHilbert.R
\name{norm.BayesHilbert}
\alias{norm.BayesHilbert}
\title{norm on tangent space at p for Bayes-Hilbert data}
\usage{
norm.BayesHilbert(u, p = NULL)
}
\description{
norm on tangent space at p for Bayes-Hilbert data
}
