% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_functional.R
\name{norm.functional}
\alias{norm.functional}
\title{norm on tangent space at p for functional data}
\usage{
norm.functional(u, p = NULL)
}
\description{
norm on tangent space at p for functional data
}
