% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{norm.manifold}
\alias{norm.manifold}
\title{norm on the tangent space at \eqn{p}}
\usage{
norm.manifold(u, p, space = "Euclid")
}
\arguments{
\item{u}{a \eqn{n}-by-\eqn{m'} matrix. Each row is a tangent vector.}

\item{p}{an \eqn{m} vector of the base point on tangent space}

\item{space}{a name of space, one of 'Euclid', 'simplex', 'sphere', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n} vector of norm. Each entry is the norm of u_i at p
}
\description{
norm on the tangent space at \eqn{p}
}
