% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensorProduct.R
\name{norm.tensor}
\alias{norm.tensor}
\title{Norm on the Tensor Product Space}
\usage{
norm.tensor(x)
}
\arguments{
\item{x}{an element in \eqn{\mathbb{H}_1\otimes \mathbb{H}_2}, created by \code{\link{make.tensor}}.}
}
\value{
the norm of x.
}
\description{
Computes the inner product of an element in \eqn{\mathbb{H}_1\otimes \mathbb{H}_2}.
}
