% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_Wasserstein.R
\name{norm.Wasserstein}
\alias{norm.Wasserstein}
\title{norm on tangent space at p for Wasserstein data}
\usage{
norm.Wasserstein(u, p = NULL)
}
\description{
norm on tangent space at p for Wasserstein data
}
