% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensorProduct.R
\name{operator.tensor}
\alias{operator.tensor}
\title{Compute operator maps from \eqn{H_1} to \eqn{H_2}}
\usage{
operator.tensor(x, y)
}
\arguments{
\item{x}{a tensor product object, created by \code{\link{make.tensor}}.}

\item{y}{an \eqn{n}-by-\eqn{n_1} matrix. Each row is an element in \eqn{H_1}.}
}
\value{
an \eqn{n}-by-\eqn{n_2} matrix. Each row is an element in \eqn{H_2}.
}
\description{
Compute operator maps from \eqn{H_1} to \eqn{H_2}
}
