% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_list.R
\name{PCA.manifold.list}
\alias{PCA.manifold.list}
\title{Principal component analysis for manifold-valued data,}
\usage{
PCA.manifold.list(Xdata)
}
\arguments{
\item{Xdata}{a list of manifold-valued data.
\describe{
      \item{j}{A \eqn{p} list of generated data. Each jth element is an \eqn{n}-by-\eqn{dim_j} matrix.}
      \item{Xspaces}{A \eqn{p} vector of underlying spaces of \eqn{X_j}.}
      \item{p}{A number of \eqn{X_j}.}
}}
}
\value{
a \eqn{p} list of PCA.manifold object, see \code{\link{PCA.manifold}}.
}
\description{
PCA fitting for a list of manifold-valued data.
}
