% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_AffInv.R, R/Geometry_SPD_LogEuclid.R,
%   R/Geometry_Wasserstein.R, R/Geometry_functional.R, R/Geometry_simplex.R,
%   R/Geometry_sphere.R
\name{PCA.manifold}
\alias{PCA.manifold}
\alias{PCA.BayesHilbert}
\alias{PCA.Euclid}
\alias{PCA.SPD.AffInv}
\alias{PCA.SPD.LogEuclid}
\alias{PCA.Wasserstein}
\alias{PCA.functional}
\alias{PCA.simplex}
\alias{PCA.sphere}
\title{Principal component analysis for manifold-valued data.}
\usage{
PCA.manifold(X, space = "Euclid", alpha = 0.95)

PCA.BayesHilbert(X, alpha = 0.95)

PCA.Euclid(X, alpha = 0.95)

PCA.SPD.AffInv(X, alpha = 0.95)

PCA.SPD.LogEuclid(X, alpha = 0.95)

PCA.Wasserstein(X, alpha = 0.95)

PCA.functional(X, alpha = 0.95)

PCA.simplex(X, alpha = 0.95)

PCA.sphere(X, alpha = 0.95)
}
\arguments{
\item{X}{an \eqn{n\times m} matrix where each row \eqn{X_i} is a point in \eqn{\mathcal{M}}.}

\item{space}{the name of space, which must be one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.}

\item{alpha}{a truncation parameter of the number of basis vectors, used only for infinite dimensional \eqn{\mathcal{M}}. Selects the first index where the cumulative variance is equal to or greater than \eqn{\alpha} of the total variance.}
}
\value{
a 'PCA.manifold' object with the following components:
\describe{
      \item{space}{the name of the underlying space \eqn{\mathcal{M}}.}
      \item{values}{an \eqn{m} vector of eigenvalues.}
      \item{vectors}{a \eqn{K\times m} matrix where each row is a corresponding orthonormal basis.}
      \item{mu}{the Frechet mean of \eqn{X}.}
      \item{dim}{a number of eigenvectors \eqn{K}}
      \item{...}{additional arguments passed into specific methods.}
}
}
\description{
Performs the Principal component analysis (spectral decomposition) for a random variable \eqn{X} taking values in the manifold \eqn{\mathcal{M}}.
}
\section{Functions}{
\itemize{
\item \code{PCA.BayesHilbert()}: Method

\item \code{PCA.Euclid()}: Method

\item \code{PCA.SPD.AffInv()}: Method

\item \code{PCA.SPD.LogEuclid()}: Method

\item \code{PCA.Wasserstein()}: Method

\item \code{PCA.functional()}: Method

\item \code{PCA.simplex()}: Method

\item \code{PCA.sphere()}: Method

}}
