% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM.R
\name{predict.AM}
\alias{predict.AM}
\title{Prediction for Additive Models.}
\usage{
\method{predict}{AM}(object, Xorgnew)
}
\arguments{
\item{object}{an \code{\link{AM}} object.}
}
\value{
an \eqn{n'\times m} matrix of predicted values \eqn{\hat{Y}_{new}}.
}
\description{
Predict \eqn{\hat{Y}_{new}} for the given \eqn{X_{new}} using an \code{\link{AM}} object.
}
