% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM.R
\name{predict.GLM}
\alias{predict.GLM}
\title{Prediction for generalized Linear Models}
\usage{
\method{predict}{GLM}(object, Xnew, is.inv.link = TRUE)
}
\arguments{
\item{object}{an \code{\link{GLM}} object.}

\item{Xnew}{a list of new observations, see the "Xdata" argument in \code{\link{PCA.manifold.list}}.}
}
\value{
an \eqn{n'\times m} matrix of predicted values \eqn{\hat{Y}_{new}}.
}
\description{
Predict \eqn{\hat{Y}_{new}} for the given \eqn{X_{new}} using an \code{\link{GLM}} object.
}
