% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM.R
\name{predict.LM}
\alias{predict.LM}
\title{Prediction for Hilbert-Schmidt Linear Models}
\usage{
\method{predict}{LM}(object, Xnew)
}
\arguments{
\item{object}{an \code{\link{LM}} object.}

\item{Xnew}{a list of new observations, see the "Xdata" argument in \code{\link{PCA.manifold.list}}.}
}
\value{
an \eqn{n'\times m} matrix of predicted values \eqn{\hat{Y}_{new}}.
}
\description{
Predict \eqn{\hat{Y}_{new}} for the given \eqn{X_{new}} using an \code{\link{LM}} object.
}
