% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{predict.LR}
\alias{predict.LR}
\title{Prediction function for a LR object}
\usage{
\method{predict}{LR}(object, Xnew)
}
\arguments{
\item{object}{a \code{\link{LR}} class object.}

\item{Xnew}{a \eqn{p} list of new observations.}
}
\value{
an \eqn{n'\times m} matrix of \eqn{\hat{Y}_{new}}.
}
\description{
Predict \eqn{\hat{Y}_{new}} for given \eqn{X_{new}}.
}
