% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_list.R
\name{predict.PCA.manifold.list}
\alias{predict.PCA.manifold.list}
\title{Prediction of Score Matrices for a List of Manifold-valued Data}
\usage{
\method{predict}{PCA.manifold.list}(object, Xdatanew)
}
\arguments{
\item{object}{a \code{\link{PCA.manifold.list}} object.}

\item{Xnew}{a new list of manifold-valued covariates, see the "Xdata" argument in \code{\link{PCA.manifold.list}}.}
}
\value{
a \eqn{p} list of score matrices.
}
\description{
Computes a list of score matrices for a list of the manifold-valued data based on \code{\link{PCA.manifold.list}}.
}
