% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_AffInv.R, R/Geometry_SPD_LogEuclid.R,
%   R/Geometry_Wasserstein.R, R/Geometry_functional.R, R/Geometry_simplex.R,
%   R/Geometry_sphere.R
\name{predict.PCA.manifold}
\alias{predict.PCA.manifold}
\alias{predict.PCA.BayesHilbert}
\alias{predict.PCA.Euclid}
\alias{predict.PCA.SPD.AffInv}
\alias{predict.PCA.SPD.LogEuclid}
\alias{predict.PCA.Wasserstein}
\alias{predict.PCA.functional}
\alias{predict.PCA.simplex}
\alias{predict.PCA.sphere}
\title{Prediction of a score matrix of manifold-valued data.}
\usage{
\method{predict}{PCA.manifold}(object, Xnew)

\method{predict}{PCA.BayesHilbert}(object, Xnew)

\method{predict}{PCA.Euclid}(object, Xnew)

\method{predict}{PCA.SPD.AffInv}(object, Xnew)

\method{predict}{PCA.SPD.LogEuclid}(object, Xnew)

\method{predict}{PCA.Wasserstein}(object, Xnew)

\method{predict}{PCA.functional}(object, Xnew)

\method{predict}{PCA.simplex}(object, Xnew)

\method{predict}{PCA.sphere}(object, Xnew)
}
\arguments{
\item{object}{a \code{\link{PCA.manifold}} object.}

\item{Xnew}{an \eqn{n\times m} matrix where each row is a point in \eqn{\mathcal{M}}.}
}
\value{
an \eqn{n\times m'} score matrix.
}
\description{
Compute a score matrix of the manifold-valued data using the principal component analysis (spectral decomposition) results made by \code{\link{PCA.manaifold}}.
}
\section{Functions}{
\itemize{
\item \code{predict(PCA.BayesHilbert)}: Method

\item \code{predict(PCA.Euclid)}: Method

\item \code{predict(PCA.SPD.AffInv)}: Method

\item \code{predict(PCA.SPD.LogEuclid)}: Method

\item \code{predict(PCA.Wasserstein)}: Method

\item \code{predict(PCA.functional)}: Method

\item \code{predict(PCA.simplex)}: Method

\item \code{predict(PCA.sphere)}: Method

}}
