% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QM.R
\name{predict.QM}
\alias{predict.QM}
\title{Prediction for quantile Linear Models}
\usage{
\method{predict}{QM}(object, Xnew)
}
\arguments{
\item{object}{an \code{\link{QM}} object.}

\item{Xnew}{a new list of covariates with the following components (see also \code{\link{covariates.generate}}):
\describe{
      \item{j}{a \eqn{p} list of manifold-valued covariates, where each \eqn{j}th element is an \eqn{n'\times T_j} matrix.}
      \item{spaces}{a \eqn{p} vector of the underlying spaces \eqn{\mathcal{M}_j} of \eqn{X_j}, see \code{\link{Check.manifold}}.}
      \item{p}{the number of \eqn{X_j}.}
}}
}
\value{
an \eqn{n'\times m} matrix of predicted values \eqn{\hat{Q}_{Y|X}(X_{new})}.
}
\description{
Predict \eqn{\hat{Q}_{Y|X}(X_{new})} for the given \eqn{X_{new}} using an \code{\link{QM}} object.
}
