% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBF_utils.R
\name{predict.Transform.Score}
\alias{predict.Transform.Score}
\title{Prediction of Transform Score Matrices}
\usage{
\method{predict}{Transform.Score}(object, score.list)
}
\arguments{
\item{score.list}{a \eqn{p} list of score matrices to be transformed.}

\item{a}{\code{\link{Transform.Score}} object.}
}
\value{
a \eqn{p} list of transformed score matrices.
}
\description{
Transforms given score matrices into the interval \eqn{[0,1]} using a \code{\link{Transform.Score}} object.
}
