% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation_QM.R
\name{QM.error.generate}
\alias{QM.error.generate}
\title{Generate a random error \eqn{\varepsilon} under quantile linear regression settings}
\usage{
QM.error.generate(
  n,
  Xdata,
  tau = 0.5,
  error.std = 1,
  error.var.type = "homoscedastic",
  error.type = "normal",
  df = 3
)
}
\arguments{
\item{n}{the number of data points.}

\item{Xdata}{a list of covariates generated by \code{\link{QM.covariates.generate}}, only used for cases where error.var.type=='heterogeneous'}

\item{tau}{the quantile of Y given X (default: 0.5).}

\item{error.std}{a standard deviation parameter for \eqn{\varepsilon}.}

\item{error.var.type}{a type of error distribution, one of 'homoscedastic' or 'heteroscedastic'.}

\item{error.type}{a distribution of \eqn{\varepsilon}, one of 'normal', 't' and 'Cauchy'.}

\item{df}{degrees of freedom (only used for error.type=='t')}
}
\value{
an \eqn{n} vector of random errors.
}
\description{
generate a random error \eqn{\varepsilon}, which follows one of normal, t and Cauchy distributions.
}
