% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_BayesHilbert.R
\name{RieExp.BayesHilbert}
\alias{RieExp.BayesHilbert}
\title{Riemannian exponential map for Bayes-Hilbert data}
\usage{
RieExp.BayesHilbert(p, u)
}
\description{
Riemannian exponential map for Bayes-Hilbert data
}
