% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_functional.R
\name{RieExp.functional}
\alias{RieExp.functional}
\title{Riemannian exponential map for functional data}
\usage{
RieExp.functional(p, u)
}
\description{
Riemannian exponential map for functional data
}
