% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{RieExp.manifold}
\alias{RieExp.manifold}
\title{Riemannian exponential map}
\usage{
RieExp.manifold(p, u, space = "Euclid")
}
\arguments{
\item{p}{an \eqn{m} vector of the base point on tangent space}

\item{u}{an \eqn{n}-by-\eqn{m'} matrix. Each row is a tangent vector.}

\item{space}{a name of space, one of 'Euclid', 'simplex', 'sphere', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n}-by-\eqn{m} matrix. Each row is a point on the Riemannian metric spaces.
}
\description{
Riemannian exponential map
}
