% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_Wasserstein.R
\name{RieExp.Wasserstein}
\alias{RieExp.Wasserstein}
\title{Riemannian exponential map for Wasserstein data}
\usage{
RieExp.Wasserstein(p, u)
}
\description{
Riemannian exponential map for Wasserstein data
}
