% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_BayesHilbert.R
\name{RieLog.BayesHilbert}
\alias{RieLog.BayesHilbert}
\title{Riemannian logarithmic map for Bayes-Hilbert data}
\usage{
RieLog.BayesHilbert(p, q)
}
\description{
Riemannian logarithmic map for Bayes-Hilbert data
}
