% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_functional.R
\name{RieLog.functional}
\alias{RieLog.functional}
\title{Riemannian logarithmic map for functional data}
\usage{
RieLog.functional(p, q)
}
\description{
Riemannian logarithmic map for functional data
}
