% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{RieLog.manifold}
\alias{RieLog.manifold}
\title{Riemannian logarithmic map on the manifold}
\usage{
RieLog.manifold(p, q, space = "Euclid")
}
\arguments{
\item{p, q}{\eqn{m} vectors or \eqn{n\times m} matrices, where each row is a point on \eqn{\mathcal{M}}.}

\item{space}{the name of space, which must be one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n\times m} matrix of Riemannian logarithmics from  corresponding rows of \eqn{p} to those of \eqn{q}.
}
\description{
Compute the Riemnnain logarithmic map on the manifold \eqn{\mathcal{M}}.
}
