% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_Wasserstein.R
\name{RieLog.Wasserstein}
\alias{RieLog.Wasserstein}
\title{Riemannian logarithmic map for Wasserstein data}
\usage{
RieLog.Wasserstein(p, q)
}
\description{
Riemannian logarithmic map for Wasserstein data
}
