% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBF_comp.R
\name{SBF.preprocessing}
\alias{SBF.preprocessing}
\title{SBF preprocessing}
\usage{
SBF.preprocessing(
  X,
  LogY,
  bandwidths,
  degree = 0,
  ngrid = 51,
  Kdenom_method = "numeric"
)
}
\arguments{
\item{X}{an \eqn{n\times p} score matrix.}

\item{LogY}{an \eqn{n\times m} response matrix.}

\item{bandwidths}{a \eqn{p} vector of bandwidths.}

\item{degree}{the degree of the local polynomial fitting. Options are: 0 for Nadaraya-Watson (default), 1 for local-linear, etc.}

\item{ngrid}{the number of grid points for evaluating the KDE (default: 51).}

\item{Kdenom_method}{a method used for the denominator in the normalized kernel calculation.}
}
\value{
a \code{SBF.comp} object with the following components: 
\describe{
      \item{tildem}{a \eqn{p} list of one-dimensional KDE where each element is (ngrid,degree+1,m) array.}
      \item{kde.1d}{a \eqn{p} list of \eqn{\tilde{m}_j} where each element is (one-dimensional KDE where each element is (ngrid,degree+1,degree+1) array.}
      \item{proj}{a \eqn{(p\times p\times ngrid\times ngrid) \times (degree+1) \times (degree+1)}-dimensional array.}
      \item{bandwidths}{a \eqn{p} vector of bandwidths.}
      \item{grids}{grid points.}
      \item{weights}{a vector of weights for fast computing numerical integration.}
}
}
\description{
Computes the pre-requisite components for smooth backfitting (SBF).
It uses \code{\link{KDE}} for density estimation.
}
