% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation.R
\name{tensor.beta.generate.each}
\alias{tensor.beta.generate.each}
\title{Generate a Hilbert-Schmidt operator between tangent spaces of \eqn{\mathcal{M}_X} and \eqn{\mathcal{M}_Y}}
\usage{
tensor.beta.generate.each(Xspace, Yspace, Xmu, Ymu, Xdim, Ydim, beta.norm = 1)
}
\arguments{
\item{Xspace}{the name of the underlying space \eqn{\mathcal{M}_X} of \eqn{X}.}

\item{Yspace}{the name of the underlying space \eqn{\mathcal{M}_Y} of \eqn{Y}.}

\item{Xmu}{the Frechet mean of \eqn{X_j}.}

\item{Ymu}{the Frechet mean of \eqn{Y}.}

\item{Xdim}{the intrinsic dimension of \eqn{\mathcal{M}_X}.}

\item{Ydim}{the intrinsic dimension of \eqn{\mathcal{M}_Y}.}

\item{beta.norm}{the tensor norm, with a default value of 1.}
}
\value{
a Hilbert-Schmidt operators (an tensor element) in \eqn{T_{\mu_X}\mathcal{M}_X \otimes T_{\mu_Y}\mathcal{M}_Y} generated by \code{\link{make.tensor}}.
}
\description{
Generate a Hilbert-Schmidt operator \eqn{\mathfrak{B}: T_{\mu_X}\mathcal{M}_X \to T_{\mu_Y}\mathcal{M}_Y}, where \eqn{\mu_X} and \eqn{\mu_Y} are the Frechet means of \eqn{X} and \eqn{Y}, respectively.
This operator can be identified as an element in the tensor product space \eqn{T_{\mu_X}\mathcal{M}_X \otimes T_{\mu_Y}\mathcal{M}_Y}.
}
