% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation.R
\name{tensor.beta.generate}
\alias{tensor.beta.generate}
\title{Generate a list of Hilbert-SChmidt operators}
\usage{
tensor.beta.generate(
  Xspaces,
  Yspace,
  Xdims,
  Ydim,
  proper.indices = NULL,
  beta.norm = 1
)
}
\arguments{
\item{Xspaces}{a \eqn{p} vector of the names of the underlying spaces \eqn{\mathcal{M}_j} of \eqn{X_j}.}

\item{Yspace}{the name of the underlying space \eqn{\mathcal{M}_Y} of \eqn{Y}.}

\item{Xdims}{a \eqn{p} vector of the intrinsic dimension of \eqn{\mathcal{M}_j}.}

\item{Ydim}{the intrinsic dimension of \eqn{\mathcal{M}_Y}.}

\item{proper.indices}{an index set \eqn{\mathcal{S}=\{1\le j\le p : \mathfrak{B}_j\neq0\}}.}

\item{beta.norm}{a \eqn{p} vector (or a single integer) of the tensor norms, used only for \eqn{j\notin\mathcal{S}}. If an integer is provided, the same value is applied for all \eqn{X_j}.}
}
\value{
a list of Hilbert-Schmidt operators (tensor elements) containing:
   \describe{
      \item{j}{For \eqn{1\le j\le p}, the \eqn{j}th element is an \eqn{n\times D_j} matrix, with each \eqn{i}th row of \eqn{X_{ij}}.}
      \item{Xspaces}{a \eqn{p} vector of the underlying spaces \eqn{\mathcal{M}_j} of \eqn{X_j}.}
      \item{p}{the number of \eqn{X_j}.}
      \item{beta.norm}{a \eqn{p} vector of the tensor norms}
}
}
\description{
Generate a list of Hilbert-Schmidt operators \eqn{\mathfrak{B}_j: T_{\mu_j}\mathcal{M}_j \to T_{\mu_Y}\mathcal{M}_Y}, where \eqn{\mu_j} and \eqn{\mu_Y} are the Frechet means of \eqn{X_j} and \eqn{Y}, respectively.
Each operator is created by \code{\link{tensor.beta.generate.each}} and can be identified as an element in \eqn{T_{\mu_j}\mathcal{M}_j \otimes T_{\mu_Y}\mathcal{M}_Y}.
}
