% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBF_utils.R
\name{Transform.Score}
\alias{Transform.Score}
\title{Transform Score Matrices}
\usage{
Transform.Score(score.list, transform = "linear", normalize = TRUE)
}
\arguments{
\item{score.list}{a \eqn{p} list of score matrices. It is only used for computing \code{min.list}, \code{max.list} and \code{omega.sq.list}.}

\item{transform}{a method of transformation. Must be one of 'linear', 'Gaussian' or 'trigonometric' (default: 'linear').}

\item{normalize}{a parameter indicating whether to standardize each column of score matrices (default: \code{TRUE}).}
}
\value{
a \code{Transform.Score} object with the following components:
\describe{
      \item{p}{the length of list of score matrices.}
      \item{index.mat}{a \eqn{L_+^* \times 3} matrix, with each row being \eqn{(i,j,k)} such that \eqn{i\equiv (j,k)}.}
      \item{min.list}{a \eqn{p} list of minimum values of columns of (normalized) score matrices.}
      \item{max.list}{a \eqn{p} list of maximum values of columns of (normalized) score matrices.}
      \item{transform}{the method of transformation.}
      \item{normalize}{the parameter indicating whether standardize each column of score matrices.}
      \item{omega.sq.list}{a \eqn{p} list of standard deviations of each column of score matrices.}
}
}
\description{
Transforms score matrices into the interval \eqn{[0,1]}.
This function is implemented for three types of transformations:
- linear: Min-max scaling.
- Gaussian: Cumulative distribution function of the standard Gaussian distribution.
- trigonometric: Inverse tangent function.
If \code{normalize} is \code{TRUE}, the function standardizes each column of the score matrices.
}
