% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation.R
\name{Ymu.generate}
\alias{Ymu.generate}
\title{Generate the Frechet mean for \eqn{Y}}
\usage{
Ymu.generate(m, space)
}
\arguments{
\item{m}{The length of each observation. Typically, this equals the dimension, but it can differ in some cases.}

\item{space}{The name of space, see \code{\link{Check.manifold}}.}
}
\value{
The Frechet mean \eqn{\mu_Y} of \eqn{Y}, an \eqn{m} vector.
}
\description{
This code is the same as \code{\link{Xmu.generate}}.
}
