% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation_AM.R
\name{AM.data.generate}
\alias{AM.data.generate}
\title{Generate additive regression data}
\usage{
AM.data.generate(
  n,
  Xspaces,
  Yspace,
  Xdims,
  Ydim,
  proper.ind.mat,
  add.mean.norm = 1,
  Xrho = 0.5,
  Xsigma = 1,
  error.rho = 0.5,
  error.std = 1,
  ngrid = 50,
  transform = "Gaussian",
  seed = 1,
  c.add.mean = NULL
)
}
\arguments{
\item{proper.ind.mat}{a \eqn{s\times 2} index matrix such that \eqn{\mathcal{S}=\{(j,k) : m_{jk}^*\neq0\}}.}

\item{add.mean.norm}{a \eqn{s} vector (or a single integer) of \eqn{\|m_{jk}^*\|_f}, used only for \eqn{(j,k)\in\mathcal{S}}. If an integer is provided, the same value is applied to all \eqn{(j,k)\in\mathcal{S}}.}

\item{seed}{a random seed, which must be greater than 1.}

\item{c.add.mean}{a list of pre-computed parameters to ensujre that \eqn{\E\|m_{jk}^*(\xi_{jk})\|^2 = (\text{add.mean.norm[j]})^2}, only computed if not provided.}
}
\value{
a list of data containing:
   \describe{
      \item{X}{a list of manifold-valued covariates, see \code{\link{covariates.generate}}.}
      \item{Y}{an \eqn{n\times m} matrix of manifold-valued responses.}
      \item{p}{the number of \eqn{X_j}.}
      \item{beta}{a \eqn{p} list of Hilbert-Schmidt operators, see \code{\link{tensor.beta.generate}}.}
      \item{Ymu}{the Frechet mean \eqn{\mu_Y} of \eqn{Y}.}
      \item{Xmu}{a \eqn{p} list of the Frechet means \eqn{\mu_j} of \eqn{X_j}.}
      \item{LogY}{the Riemannian logarithmic transformation \eqn{Log_{\mu_Y}Y}.}
      \item{LogX}{a \eqn{p} list of \eqn{Log_{\mu_j}X_j}, the Riemannian logarithmic transformations of \eqn{X_j}.}
      \item{Xbeta.each}{a \eqn{p} list of \eqn{\mathfrak{B}_j(Log_{\mu_j}X_j)}.}
      \item{Xbeta}{the sum of Xbeta.each, i.e., \eqn{\sum_{j=1}^p\mathfrak{B}_j(Log_{\mu_j}X_j)}.}
      \item{ExpXbeta}{the Riemannian exponential of Xbeta, i.e., \eqn{Exp_{\mu_Y}\Big(\sum_{j=1}^p\mathfrak{B}_j(Log_{\mu_j}X_j)\Big)}.}
      \item{...}{other input parameters.}
}
}
\description{
Generate additive regression data.
The seed for generating additive mean functiions \eqn{m_{jk}^*} is fixed as zero, ensuring that each \eqn{m_{jk}^*} is the same across simulations.
}
