% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_Euclid.R,
%   R/Geometry_simplex.R, R/Geometry_sphere.R
\name{basis.manifold}
\alias{basis.manifold}
\alias{basis.Euclid}
\alias{basis.simplex}
\alias{basis.sphere}
\title{Orthonormal basis on the tangent space of the manifold}
\usage{
basis.manifold(p, dim = 50, space = "Euclid")

basis.Euclid(p, dim = NULL)

basis.simplex(p, dim = NULL)

basis.sphere(p, dim = NULL)
}
\arguments{
\item{p}{an \eqn{m} vector representing the base point in \eqn{\mathcal{M}}.}

\item{dim}{the number of orthonormal basis, only used for infinite-dimensional \eqn{\mathcal{M}}.}

\item{space}{the name of space, which must be one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n\times m} matrix where each row is an orthonormal basis of the tangent space \eqn{T_p\mathcal{M}}.
}
\description{
Compute the orthonormal basis of the tangent space \eqn{T_p\mathcal{M}}.
}
\section{Functions}{
\itemize{
\item \code{basis.Euclid()}: Method

\item \code{basis.simplex()}: Method

\item \code{basis.sphere()}: Method

}}
