% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{basis.manifold}
\alias{basis.manifold}
\title{Basis on the tangent space at p}
\usage{
basis.manifold(p, dim = 50, space = "Euclid")
}
\arguments{
\item{p}{an \eqn{m} vector of the point on Riemannian metric spaces.}

\item{dim}{a number of basis, only used for infinite-dimensional Riemannian metric spaces.}

\item{space}{a name of space, one of 'Euclid', 'simplex', 'sphere', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
a \eqn{n}-by-\eqn{dim} matrix with dimension. Each row is an orthonormal basis of the tangent space at p.
}
\description{
Basis on the tangent space at p
}
