% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_BayesHilbert.R,
%   R/Geometry_SPD_LogEuclid.R, R/Geometry_Wasserstein.R,
%   R/Geometry_functional.R
\name{basis.BayesHilbert}
\alias{basis.BayesHilbert}
\alias{basis.SPD.LogEuclid}
\alias{basis.Wasserstein}
\alias{basis.functional}
\title{Basis on the tangent space at the point for Bayes-Hilbert data}
\usage{
basis.BayesHilbert(p, dim = 50)

basis.SPD.LogEuclid(p, dim = 50)

basis.Wasserstein(p, dim = 50)

basis.functional(p, dim = 50)
}
\description{
Basis on the tangent space at the point for Bayes-Hilbert data

Basis on the tangent space at the point for SPD data with the Log-Euclidean metric

Basis on the tangent space at the point for Wasserstein data

Basis on the tangent space at the point for functional data
}
\section{Functions}{
\itemize{
\item \code{basis.BayesHilbert()}: Method

\item \code{basis.SPD.LogEuclid()}: Method

\item \code{basis.Wasserstein()}: Method

\item \code{basis.functional()}: Method

}}
