% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{Check.manifold}
\alias{Check.manifold}
\title{Basic geometry structure on manifolds}
\usage{
Check.manifold(space)
}
\arguments{
\item{space}{a name of space, one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\description{
Basic:
inner product, norm in tangent spaces, distance
exponential and logarithmic maps
}
\details{
For PCA:
Frechet mean, PCA

Check whether manifolds or not
}
