% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_simplex.R
\name{clr}
\alias{clr}
\title{Centered log-ratio transform}
\usage{
clr(x)
}
\arguments{
\item{x}{an \eqn{n\times m} matrix of simplex- or Bayes-Hilbert space-valued data.}
}
\value{
an \eqn{n\times m} matrix of clr-transformed data.
}
\description{
Compute the centered log-ratio transform for compositional or density data.
Specifically, for each \eqn{x=(x_i)\in\mathbb{R}^m}, the transformation is give by
\deqn{\text{clr}(x)_i = \log(x_i) - \frac{1}{m}\sum_{j}\log(x_j)}.
}
