% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_data_generation.R
\name{covariates.generate.each}
\alias{covariates.generate.each}
\title{Generate each manifold-valued covariates}
\usage{
covariates.generate.each(Xi, dim, space = "Euclid")
}
\arguments{
\item{Xi}{an \eqn{n\times m'} score matrix.}

\item{dim}{a dimension of \eqn{X}.}

\item{space}{the underlying spaces, see \code{\link{Check.manifold}}}
}
\value{
an \eqn{n\times m} matrix with dimension. Each row is an element of the manifold.
}
\description{
Generate each manifold-valued covariates
}
