% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_data_generation.R
\name{covariates.generate}
\alias{covariates.generate}
\title{Generate a list of Riemannian metric space-valued covariates.}
\usage{
covariates.generate(n, Xspaces, dims, Xrho = 0.5, Xsigma = 1)
}
\arguments{
\item{n}{a number of data}

\item{Xspaces}{a \eqn{p} vector of underlying spaces of \eqn{X_j}.}

\item{dims}{a \eqn{p} vector of dimension of \eqn{X_j}.}

\item{Xrho}{a correlation parameter which ranges from -1 to 1.}

\item{Xsigma}{a correlation parameter which ranges from -1 to 1.}
}
\value{
a list of generated data
   \describe{
      \item{j}{A \eqn{p} list of generated data. Each jth element is an \eqn{n}-by-\eqn{dim_j} matrix.}
      \item{Xspaces}{A \eqn{p} vector of underlying spaces of \eqn{X_j}.}
      \item{p}{A number of \eqn{X_j}.}
}
}
\description{
Generate a list of Riemannian metric space-valued covariates.
}
