% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation_LM.R
\name{covariates.generate}
\alias{covariates.generate}
\title{Generate a list of manifold-valued covariates.}
\usage{
covariates.generate(n, Xspaces, Xmu.list, Xdims, Xrho = 0.5, Xsigma = 1, a = 1)
}
\arguments{
\item{n}{the number of data points.}

\item{Xspaces}{a \eqn{p} vector specifying the underlying spaces of \eqn{X_j}.}

\item{Xmu.list}{a \eqn{p} list of the Frechet means of \eqn{X_j}.}

\item{Xdims}{a \eqn{p} vector specifying the dimensions of \eqn{X_j}.}

\item{Xrho}{a correlation parameter ranging from -1 to 1, with a default value of 0.5.}

\item{Xsigma}{a common standard deviation parameter, with a default value of 1.}

\item{a}{a parameter such that \eqn{\text{Var}(\xi_{jk})\asymp k^{-2a}}.}
}
\value{
a list of data containing:
   \describe{
      \item{j}{a \eqn{p}-list of generated data, where each \eqn{j}th element is an \eqn{n\times D_j} matrix.}
      \item{Xspaces}{a \eqn{p} vector specifying the underlying spaces of \eqn{X_j}, see \code{\link{Check.manifold}}.}
      \item{p}{the number of \eqn{X_j}.}
}
}
\description{
Generate a list of manifold-valued covariates.
If the underlying space of \eqn{X_j} is finite-dimensional, the dimension of \eqn{X_j} is set to \eqn{D_j}.
If the underlying space of \eqn{X_j} is inifite-dimensional, the dimension of \eqn{X_j} is set to 100.
}
