% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation.R
\name{covariates.generate.real}
\alias{covariates.generate.real}
\title{Generate real covariates}
\usage{
covariates.generate.real(n, p, Zrho, Zsigma)
}
\arguments{
\item{n}{a number of data.}

\item{p}{a number of covariates.}

\item{Zrho}{a correlation parameter which ranges from -1 to 1.}

\item{Zsigma}{a variance parameter, default 1.}
}
\value{
an \eqn{n\times p} covariate matrix.
}
\description{
Generate real covariates
}
