% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation.R
\name{covariates.generate.real}
\alias{covariates.generate.real}
\title{Generate real covariates}
\usage{
covariates.generate.real(n, p, Zrho, Zsigma)
}
\arguments{
\item{n}{the number of data points.}

\item{p}{the number of covariates.}

\item{Zrho}{a correlation parameter ranging from -1 to 1.}

\item{Zsigma}{a variance parameter, with a default value of 1.}
}
\value{
an \eqn{n\times p} covariate matrix.
}
\description{
Generate an \eqn{n\times p} matrix \eqn{\bm{X}=(X_{ij})}, where each \eqn{X_{ij}} represents a covariate
}
