% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{dist.manifold}
\alias{dist.manifold}
\title{Geodesic distance on the manfold}
\usage{
dist.manifold(p, q, space = "Euclid")
}
\arguments{
\item{p, q}{\eqn{m} vectors or \eqn{n}-by-\eqn{m} matrices. Each row is a point on the Riemannian metric spaces.}

\item{space}{a name of space, one of 'Euclid', 'simplex', 'sphere', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n} vector of geodesic distance. Each entry is a distance between \eqn{p_i} and \eqn{q_i}.
}
\description{
Geodesic distance on the manfold
}
