% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_AffInv.R, R/Geometry_SPD_LogEuclid.R,
%   R/Geometry_Wasserstein.R, R/Geometry_functional.R, R/Geometry_simplex.R,
%   R/Geometry_sphere.R
\name{dist.manifold}
\alias{dist.manifold}
\alias{dist.BayesHilbert}
\alias{dist.Euclid}
\alias{dist.SPD.AffInv}
\alias{dist.SPD.LogEuclid}
\alias{dist.Wasserstein}
\alias{dist.functional}
\alias{dist.simplex}
\alias{dist.sphere}
\title{Geodesic distance on the manfold}
\usage{
dist.manifold(p, q, space = "Euclid")

dist.BayesHilbert(p, q)

dist.Euclid(p, q)

dist.SPD.AffInv(p, q)

dist.SPD.LogEuclid(p, q)

dist.Wasserstein(p, q)

dist.functional(p, q)

dist.simplex(p, q)

dist.sphere(p, q)
}
\arguments{
\item{p, q}{\eqn{m} vectors or \eqn{n\times m} matrices. Each row is a point on the Riemannian metric spaces.}

\item{space}{a name of space, one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n} vector of geodesic distance. Each entry is a distance between \eqn{p_i} and \eqn{q_i}.
}
\description{
Geodesic distance on the manfold

Geodesic distance for Bayes-Hilbert data

Geodesic distance for Euclid data

Geodesic distance for SPD data with the Affine-invariant metric

Geodesic distance for SPD data with the Log-Euclidean metric

Geodesic distance for Wasserstein data

Geodesic distance for functional data

Geodesic distance for simplex data

Geodesic distance for sphere data
}
\section{Functions}{
\itemize{
\item \code{dist.BayesHilbert()}: Method

\item \code{dist.Euclid()}: Method

\item \code{dist.SPD.AffInv()}: Method

\item \code{dist.SPD.LogEuclid()}: Method

\item \code{dist.Wasserstein()}: Method

\item \code{dist.functional()}: Method

\item \code{dist.simplex()}: Method

\item \code{dist.sphere()}: Method

}}
