% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation.R
\name{error.generate}
\alias{error.generate}
\title{Generate a random error \eqn{\varepsilon} of \eqn{Y}}
\usage{
error.generate(n, Yspace, Ymu, Ydim, error.rho = 0.5, error.std = 1)
}
\arguments{
\item{n}{the number of data points.}

\item{Yspace}{the underlying space of \eqn{\mathcal{M}_Y}.}

\item{Ymu}{the Frechet mean of \eqn{Y}.}

\item{Ydim}{the intrinsic dimension of \eqn{\mathcal{M}_Y}.}

\item{error.rho}{a correlation parameter for \eqn{\varepsilon}.}

\item{error.std}{a standard deviation parameter for \eqn{\varepsilon}.}
}
\value{
an \eqn{n\times m} matrix of random errors.
}
\description{
For a random element \eqn{Y} taking values in a manifold \eqn{\mathcal{M}_Y} with the Frechet mean \eqn{\mu_Y}, 
generate a random error \eqn{\varepsilon} taking values in \eqn{T_{\mu_Y}\mathcal{M}_Y}.
}
