% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_LogEuclid.R, R/Geometry_Wasserstein.R,
%   R/Geometry_functional.R, R/Geometry_simplex.R, R/Geometry_sphere.R
\name{FrechetMean.manifold}
\alias{FrechetMean.manifold}
\alias{FrechetMean.BayesHilbert}
\alias{FrechetMean.Euclid}
\alias{FrechetMean.SPD.LogEuclid}
\alias{FrechetMean.Wasserstein}
\alias{FrechetMean.functional}
\alias{FrechetMean.simplex}
\alias{FrechetMean.sphere}
\title{Frechet mean on the manifold}
\usage{
FrechetMean.manifold(X, space = "Euclid")

FrechetMean.BayesHilbert(X)

FrechetMean.Euclid(X)

FrechetMean.SPD.LogEuclid(X)

FrechetMean.Wasserstein(X)

FrechetMean.functional(X)

FrechetMean.simplex(X)

FrechetMean.sphere(X)
}
\arguments{
\item{X}{a \eqn{n\times m} matrix. Each row is a point on the Riemannian metric spaces.}

\item{space}{a name of space, one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{m} vector
}
\description{
Frechet mean on the manifold

Frechet mean for Bayes-Hilbert data

Frechet mean for Euclid data

Frechet mean for SPD.LogEuclid data

Frechet mean for Wasserstein data

Frechet mean for functional data

Frechet mean for simplex data

Frechet mean for sphere data
}
\section{Functions}{
\itemize{
\item \code{FrechetMean.BayesHilbert()}: Method

\item \code{FrechetMean.Euclid()}: Method

\item \code{FrechetMean.SPD.LogEuclid()}: Method

\item \code{FrechetMean.Wasserstein()}: Method

\item \code{FrechetMean.functional()}: Method

\item \code{FrechetMean.simplex()}: Method

\item \code{FrechetMean.sphere()}: Method

}}
