% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation.R
\name{GLM.data.generate}
\alias{GLM.data.generate}
\title{Generate generalized linear regression data}
\usage{
GLM.data.generate(
  n,
  Xspaces,
  Xdims,
  link = "binomial",
  Ydim = 1,
  proper.indices = NULL,
  beta.norm = 1,
  beta0.norm = 1,
  Xrho = 0.5,
  Xsigma = 1,
  seed = 1
)
}
\arguments{
\item{n}{a number of observation.}

\item{Xspaces}{a \eqn{p} vector of underlying spaces of \eqn{X_j}.}

\item{Xdims}{a \eqn{p} vector of intrinsic dimensions of \eqn{H_j}.}

\item{link}{a link function, see \code{\link{Check.link}}.}

\item{Ydim}{a dimension of \eqn{Y}. It is only used when link is multinomial.}

\item{proper.indices}{an indices of nonzero operators.}

\item{beta.norm}{a \eqn{p} vector or an integer of the tensor norm. It is only used for \eqn{j\notin}proper.indices. If it is an integer, we use the same value for all \eqn{X_j}.}

\item{beta0.norm}{an integer of the norm of \eqn{\beta_0^*}.}

\item{Xrho}{a correlation parameter for \eqn{X_j}.}

\item{Xsigma}{a variance parameter for \eqn{X_j}.}

\item{seed}{a random seed (1<=seed)}
}
\value{
a list of generated data
   \describe{
      \item{X}{a \eqn{p} list of generated covariates. Each \eqn{X_j} is an \eqn{n\times Xdim_j} matrix.}
      \item{Y}{an \eqn{n\times m} matrix of response.}
      \item{theta}{a canonical parameter.}
      \item{Ymu}{Inverse link of theta.}
      \item{beta}{a \eqn{p} list of generated operators.}
      \item{beta0}{a constant beta0 for }
      \item{error}{a generated random error.}
      \item{Xbeta.each}{a \eqn{p} list of \eqn{B_j(X_j)}.}
      \item{Xbeta}{a sum of Xbeta.each.}
      \item{...}{the other input parameters.}
}
}
\description{
Generate generalized linear regression data.
The seed for generating beta is fixed as zero, so each beta is equal in simulations.
}
