% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_oracle_CV.R
\name{GLM.oracle.GCV}
\alias{GLM.oracle.GCV}
\title{GCV for an LM.oracle function}
\usage{
GLM.oracle.GCV(
  Xorg,
  Yorg,
  Xorgnew,
  Yorgnew,
  Xdim.max.list,
  proper.indices = NULL,
  link = "binomial",
  max.cv.iter = 20,
  cv.threshold = 1e-10,
  eta = 0.001,
  max.iter = 500,
  threshold = 1e-10
)
}
\arguments{
\item{Xorg}{a list of manifold-valued covariates, see \code{\link{PCA.manifold.list}}.}

\item{Yorg}{an \eqn{n\times m} response matrix.}

\item{Xdim.max.list}{a vector of max dimension of \eqn{X_j}.}

\item{proper.indices}{a vector of indices of relevant \eqn{X_j}.}

\item{Xnew}{a new list of manifold-valued covariates.}

\item{Ynew}{a new \eqn{n'\times m} response matrix.}

\item{Yspace}{an underlying space of Yorg and Ynew.}
}
\value{
an \eqn{\code{LM}} object
   \describe{
      \item{Xdim.max.list}{a vector of Xdim.max.}
      \item{loss.list}{a list of loss for each CV step.}
      \item{runtime}{running time.}
      \item{...}{see \code{\link{LM.oracle}}.}
}
}
\description{
GCV for an LM.oracle function
}
