% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{inner.manifold}
\alias{inner.manifold}
\title{Inner product on the tangent space at \eqn{p}}
\usage{
inner.manifold(u, v, p, space = "Euclid")
}
\arguments{
\item{u, v}{\eqn{n}-by-\eqn{m'} Matrices. Each row is tangent vector.}

\item{p}{an \eqn{m} vector of the base point on tangent space}

\item{space}{a name of space, one of 'Euclid', 'simplex', 'sphere', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n} vector of inner products of each row of u and v
}
\description{
Inner product on the tangent space at \eqn{p}
}
