% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_AffInv.R, R/Geometry_SPD_LogEuclid.R,
%   R/Geometry_Wasserstein.R, R/Geometry_functional.R, R/Geometry_simplex.R,
%   R/Geometry_sphere.R
\name{inner.manifold}
\alias{inner.manifold}
\alias{inner.BayesHilbert}
\alias{inner.Euclid}
\alias{inner.SPD.AffInv}
\alias{inner.SPD.LogEuclid}
\alias{inner.Wasserstein}
\alias{inner.functional}
\alias{inner.simplex}
\alias{inner.sphere}
\title{Inner product on the tangent space at \eqn{p}}
\usage{
inner.manifold(u, v, p, space = "Euclid")

inner.BayesHilbert(u, v, p = NULL)

inner.Euclid(u, v, p = NULL)

inner.SPD.AffInv(u, v, p = NULL)

inner.SPD.LogEuclid(u, v, p = NULL)

inner.Wasserstein(u, v, p = NULL)

inner.functional(u, v, p = NULL)

inner.simplex(u, v, p = NULL)

inner.sphere(u, v, p = NULL)
}
\arguments{
\item{u, v}{\eqn{n\times m'} Matrices. Each row is tangent vector.}

\item{p}{an \eqn{m} vector of the base point on tangent space}

\item{space}{a name of space, one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n} vector of inner products of each row of u and v
}
\description{
Inner product on the tangent space at \eqn{p}

Inner product on tangent space at p for Bayes-Hilbert data

Inner product on tangent space at p for Euclid data

Inner product on tangent space at p for SPD data with the Affine-invariant metric

Inner product on tangent space at p for SPD data with the Log-Euclidean metric

Inner product on tangent space at p for Wasserstein data

Inner product on tangent space at p for functional data

Inner product on tangent space at p for simplex data

Inner product on tangent space at p for sphere data
}
\section{Functions}{
\itemize{
\item \code{inner.BayesHilbert()}: Method

\item \code{inner.Euclid()}: Method

\item \code{inner.SPD.AffInv()}: Method

\item \code{inner.SPD.LogEuclid()}: Method

\item \code{inner.Wasserstein()}: Method

\item \code{inner.functional()}: Method

\item \code{inner.simplex()}: Method

\item \code{inner.sphere()}: Method

}}
