% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensorProduct.R
\name{inner.tensor}
\alias{inner.tensor}
\title{Inner product on the Tensor Product Space}
\usage{
inner.tensor(x, y)
}
\arguments{
\item{x, y}{elements in \eqn{\mathbb{H}_1\otimes \mathbb{H}_2}, created by \code{\link{make.tensor}}.}
}
\value{
the inner product of x and y.
}
\description{
Computes the inner product of two elements in \eqn{\mathbb{H}_1\otimes \mathbb{H}_2}.
}
