% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry_BayesHilbert.R
\name{inv.clr.BayesHilbert}
\alias{inv.clr.BayesHilbert}
\title{Inverse centered log-transform for density data}
\usage{
inv.clr.BayesHilbert(x, normalizer = 1)
}
\arguments{
\item{x}{clr data, (n,m) matrix}

\item{normalizer}{normalized constant, default=1}
}
\value{
density data, (n,m) matrix
}
\description{
Inverse transform R^m to density data
inv.clr(x)_i = exp(x_i)/(sum_j exp(x_i)) * normalizer
}
