% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_oracle_CV.R
\name{LM.oracle.kfold}
\alias{LM.oracle.kfold}
\title{Kfold cross validation for an LM.oracle function}
\usage{
LM.oracle.kfold(
  Xall,
  Yall,
  Yspace,
  kfold,
  Xdim.max.list,
<<<<<<< HEAD
  proper.indices = NULL,
=======
  proper.indices,
>>>>>>> 16e5ec28b2b82a0e80cfecce213851321244bcd9
  seed = NULL
)
}
\arguments{
\item{Xall}{a list of manifold-valued covariates, see \code{\link{PCA.manifold.list}}.}

\item{Yall}{an \eqn{n\times m} response matrix.}

\item{Yspace}{an underlying space of Yorg and Ynew.}

\item{kfold}{a number of kfold CV, int>0.}

\item{Xdim.max.list}{a vector of max dimension of \eqn{X_j}.}

\item{proper.indices}{a vector of indices of relevant \eqn{X_j}.}

\item{seed}{a random seed, int>0, default: non-random (NULL).}
}
\value{
an \eqn{\code{LM}} object
   \describe{
      \item{Xdim.max.list}{a vector of Xdim.max.}
      \item{loss.list}{a list of loss for each CV step.}
      \item{runtime}{running time.}
      \item{...}{see \code{\link{LM.oracle}}.}
}
}
\description{
Kfold cross validation for an LM.oracle function
}
