% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_oracle.R
\name{LR.oracle}
\alias{LR.oracle}
\title{High-dimensional linear regression for manifold-valued responses and covariates.}
\usage{
LR.oracle(Xorg, Yorg, Yspace, Xdim.max = 100, proper.indices = NULL)
}
\arguments{
\item{Xorg}{a list of manifold-valued covariates, see \code{\link{PCA.manifold.list}}.}

\item{Yorg}{an \eqn{n\times m} response matrix.}

\item{Yspace}{an underlying space of \eqn{Y}.}

\item{Xdim.max}{a max dimension of \eqn{X_j}, real>0.}

\item{proper.indices}{a vector of indices of relevant \eqn{X_j}.}
}
\value{
an \code{\link{LR}} object.
   \describe{
      \item{pca}{an \code{\link{PCA.manifold.list}} object.}
      \item{Ymu}{an \eqn{m} vector of the Frechet mean of \eqn{Y}.}
      \item{beta}{a \eqn{P\times m} matrix of estimated beta, where \eqn{P=\sum_{j=1}^p K_j}.}
      \item{beta.each}{a \eqn{p} list of each \eqn{beta_j}.}
      \item{beta.norm}{a \eqn{p} vector of norm of each \eqn{beta_j}.}
      \item{beta.vectors}{a \eqn{p} list of corresponding bases of \eqn{X_j}. Each basis is a \eqn{K_j}-by-\eqn{T_j} matrix.}
      \item{beta.tensor}{a \eqn{p} list of tensor operators, see \code{\link{make.tensor}}.}
      \item{proper.indices}{an indices of nonzero \code{beta_j}.}
      \item{runtime}{a running time.}
      \item{...}{other input parameters.}
}
}
\description{
This function is based on a high-dimensional linear regression models with the knowledge of the nonzero indices.
}
