% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensorProduct.R
\name{make.tensor}
\alias{make.tensor}
\title{Make an element in a Tensor Product Space}
\usage{
make.tensor(element1, element2, space1, space2, mu1, mu2)
}
\arguments{
\item{element1}{an \eqn{n\times m_1} matrix, where each row is an element in \eqn{\mathbb{H}_1}.}

\item{element2}{an \eqn{n\times m_2} matrix, where each row is an element in \eqn{\mathbb{H}_2}.}

\item{space1}{the name of the underlying space \eqn{\mathbb{H}_1}.}

\item{space2}{the name of the underlying space \eqn{\mathbb{H}_2}.}

\item{mu1}{a base point of \eqn{\mathbb{H}_1}.}

\item{mu2}{a base point of \eqn{\mathbb{H}_2}.}
}
\value{
a "tensor" object with the following arguments:
   \describe{
      \item{n}{the number of elements.}
      \item{element1}{an \eqn{n\times m_1} matrix, where each row is an element in \eqn{\mathbb{H}_1}.}
      \item{element2}{an \eqn{n\times m_2} matrix, where each row is an element in \eqn{\mathbb{H}_2}.}
      \item{mu1}{a base point of \eqn{\mathbb{H}_1}.}
      \item{mu2}{a base point of \eqn{\mathbb{H}_2}.}
      \item{space1}{the name of the underlying space \eqn{\mathbb{H}_1}.}
      \item{space2}{the name of the underlying space \eqn{\mathbb{H}_2}.}
      \item{dim}{the intrinsic dimension of \eqn{\mathbb{H}_1}.}
}
}
\description{
Given two Hilbert spaces \eqn{\mathbb{H}_1} and \eqn{\mathbb{H}_2}, make an element in the tensor product space \eqn{H_1\otimes H_2}.
}
