% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_Wasserstein.R, R/Geometry_functional.R,
%   R/Geometry_simplex.R, R/Geometry_sphere.R
\name{norm.manifold}
\alias{norm.manifold}
\alias{norm.BayesHilbert}
\alias{norm.Euclid}
\alias{norm.Wasserstein}
\alias{norm.functional}
\alias{norm.simplex}
\alias{norm.sphere}
\title{norm on the tangent space at \eqn{p}}
\usage{
norm.manifold(u, p, space = "Euclid")

norm.BayesHilbert(u, p = NULL)

norm.Euclid(u, p = NULL)

norm.Wasserstein(u, p = NULL)

norm.functional(u, p = NULL)

norm.simplex(u, p = NULL)

norm.sphere(u, p = NULL)
}
\arguments{
\item{u}{a \eqn{n\times m'} matrix. Each row is a tangent vector.}

\item{p}{an \eqn{m} vector of the base point on tangent space}

\item{space}{a name of space, one of 'Euclid', 'simplex', 'sphere', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n} vector of norm. Each entry is the norm of \eqn{u_i} at p
}
\description{
norm on the tangent space at \eqn{p}

norm on tangent space at p for Bayes-Hilbert data

norm on tangent space at p for Euclid data

norm on tangent space at p for Wasserstein data

norm on tangent space at p for functional data

norm on tangent space at p for simplex data

norm on tangent space at p for sphere data
}
\section{Functions}{
\itemize{
\item \code{norm.BayesHilbert()}: Method

\item \code{norm.Euclid()}: Method

\item \code{norm.Wasserstein()}: Method

\item \code{norm.functional()}: Method

\item \code{norm.simplex()}: Method

\item \code{norm.sphere()}: Method

}}
