% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_AffInv.R, R/Geometry_SPD_LogEuclid.R,
%   R/Geometry_Wasserstein.R, R/Geometry_functional.R, R/Geometry_simplex.R,
%   R/Geometry_sphere.R
\name{norm.manifold}
\alias{norm.manifold}
\alias{norm.BayesHilbert}
\alias{norm.Euclid}
\alias{norm.SPD.AffInv}
\alias{norm.SPD.LogEuclid}
\alias{norm.Wasserstein}
\alias{norm.functional}
\alias{norm.simplex}
\alias{norm.sphere}
\title{Norm on the tangent space}
\usage{
norm.manifold(u, p, space = "Euclid")

norm.BayesHilbert(u, p = NULL)

norm.Euclid(u, p = NULL)

norm.SPD.AffInv(u, p = NULL)

norm.SPD.LogEuclid(u, p = NULL)

norm.Wasserstein(u, p = NULL)

norm.functional(u, p = NULL)

norm.simplex(u, p = NULL)

norm.sphere(u, p = NULL)
}
\arguments{
\item{u}{an \eqn{n\times m'} matrix, where each row is a tangent vector on \eqn{T_p\mathcal{M}}.}

\item{p}{an \eqn{m} vector representing the base point in \eqn{\mathcal{M}}.}

\item{space}{the name of space, which must be one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n} vector of norms for each row of \eqn{u}.
}
\description{
Compute the norm on the tangent space \eqn{T_p\mathcal{M}} of the manifold \eqn{\mathcal{M}}.
}
\section{Functions}{
\itemize{
\item \code{norm.BayesHilbert()}: Method

\item \code{norm.Euclid()}: Method

\item \code{norm.SPD.AffInv()}: Method

\item \code{norm.SPD.LogEuclid()}: Method

\item \code{norm.Wasserstein()}: Method

\item \code{norm.functional()}: Method

\item \code{norm.simplex()}: Method

\item \code{norm.sphere()}: Method

}}
