% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensorProduct.R
\name{norm.tensor}
\alias{norm.tensor}
\title{Norm of tensor product spaces}
\usage{
norm.tensor(x)
}
\arguments{
\item{x}{an element in \eqn{H_1\otimes H_2}, created by \code{\link{make.TPspace}}.}
}
\value{
a norm of x
}
\description{
Norm of tensor product spaces
}
