% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_list.R
\name{PCA.manifold.list}
\alias{PCA.manifold.list}
\title{Principal Component Analysis for a List of Manifold-valued Data.}
\usage{
PCA.manifold.list(Xdata, alpha = 0.95)
}
\arguments{
\item{Xdata}{a list of covariates with the following arguments (see also \code{\link{covariates.generate}}):
\describe{
      \item{j}{a \eqn{p} list of manifold-valued covariates, where each \eqn{j}th element is an \eqn{n\times T_j} matrix.}
      \item{spaces}{a \eqn{p} vector of the underlying spaces \eqn{\mathcal{M}_j} of \eqn{X_j}, see \code{\link{Check.manifold}}.}
      \item{p}{the number of \eqn{X_j}.}
}}

\item{alpha}{a truncation parameter of the number of basis vectors, used only for infinite dimensional \eqn{\mathcal{M}}. Selects the first index where the cumulative variance is equal to or greater than \eqn{\alpha} of the total variance.}
}
\value{
a \code{PCA.manifold.list} object with the following arguments:
\describe{
      \item{j}{a \code{\link{PCA.manifold}} object for \eqn{X_j}.}
      \item{spaces}{a \eqn{p} vector of the underlying spaces \eqn{\mathcal{M}_j} of \eqn{X_j}.}
      \item{p}{the number of \eqn{X_j}.}
}
}
\description{
Performs the principal component analysis (spectral decomposition) for a list of manifold-valued covariates.
}
