% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_list.R
\name{PCA.manifold.list}
\alias{PCA.manifold.list}
\title{Principal component analysis for manifold-valued data,}
\usage{
PCA.manifold.list(Xdata, alpha = 0.95)
}
\arguments{
\item{Xdata}{a list of manifold-valued data.
\describe{
      \item{j}{a \eqn{p} list of generated data. Each \eqn{j}th element is an \eqn{n\times T_j} matrix.}
      \item{spaces}{a \eqn{p} vector of underlying spaces of \eqn{X_j}.}
      \item{p}{a number of \eqn{X_j}.}
}}

\item{alpha}{a truncation parameter of the number of vectors, only used for infinite dimensional manifolds. Select the first index where the sum of the variances is equal to or greater than the alpha of the total.}
}
\value{
a PCA.manifold.list object.
\describe{
      \item{j}{a \code{\link{PCA.manifold}} object for \eqn{X_j}.}
      \item{spaces}{a \eqn{p} vector of underlying spaces of \eqn{X_j}.}
      \item{p}{a number of \eqn{X_j}.}
}
}
\description{
PCA fitting for a list of manifold-valued data.
}
