% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_LogEuclid.R, R/Geometry_Wasserstein.R,
%   R/Geometry_functional.R, R/Geometry_simplex.R, R/Geometry_sphere.R
\name{PCA.manifold}
\alias{PCA.manifold}
\alias{PCA.BayesHilbert}
\alias{PCA.Euclid}
\alias{PCA.SPD.LogEuclid}
\alias{PCA.Wasserstein}
\alias{PCA.functional}
\alias{PCA.simplex}
\alias{PCA.sphere}
\title{Principal component analysis for manifold-valued data.}
\usage{
PCA.manifold(X, space = "Euclid")

PCA.BayesHilbert(X)

PCA.Euclid(X)

PCA.SPD.LogEuclid(X)

PCA.Wasserstein(X)

PCA.functional(X)

PCA.simplex(X)

PCA.sphere(X)
}
\value{
a 'PCA.manifold' object
\describe{
      \item{space}{a name of the underlying space.}
      \item{values}{an \eqn{m} vector of eigenvalues.}
      \item{vectors}{a \eqn{K\times m} matrix. Each row is an orthornormal basis generated by PCA.}
      \item{mu}{the Frechet mean of X, \eqn{m} vector.}
      \item{dim}{a number of eigenvectors \eqn{K}}
      \item{...}{Passed into specific method}
}
}
\description{
Principal component analysis for manifold-valued data.

Principal component analysis for Bayes-Hilbert data

Principal component analysis for Euclid data

Principal component analysis for SPD data with the Log-Euclidean metric

Principal component analysis for Wasserstein data

Principal component analysis for functional data

Principal component analysis for simplex data

Principal component analysis for sphere data
}
\section{Functions}{
\itemize{
\item \code{PCA.BayesHilbert()}: Method

\item \code{PCA.Euclid()}: Method

\item \code{PCA.SPD.LogEuclid()}: Method

\item \code{PCA.Wasserstein()}: Method

\item \code{PCA.functional()}: Method

\item \code{PCA.simplex()}: Method

\item \code{PCA.sphere()}: Method

}}
