% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM.R
\name{predict.GLM}
\alias{predict.GLM}
\title{Prediction function for a LM object}
\usage{
\method{predict}{GLM}(object, Xnew, is.inv.link = TRUE)
}
\arguments{
\item{object}{a \code{\link{LM}} class object.}

\item{Xnew}{a \eqn{p} list of new observations.}
}
\value{
an \eqn{n'\times m} matrix of \eqn{\hat{Y}_{new}}.
}
\description{
Predict \eqn{\hat{Y}_{new}} for given \eqn{X_{new}}.
}
