% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM.R
\name{predict.LM}
\alias{predict.LM}
\title{Prediction for Hilbert-Schmidt Linear Models}
\usage{
\method{predict}{LM}(object, Xnew)
}
\arguments{
\item{object}{an \code{\link{LM}} object.}

\item{Xnew}{a new list of covariates with the following components (see also \code{\link{covariates.generate}}):
\describe{
      \item{j}{a \eqn{p} list of manifold-valued covariates, where each \eqn{j}th element is an \eqn{n'\times T_j} matrix.}
      \item{spaces}{a \eqn{p} vector of the underlying spaces \eqn{\mathcal{M}_j} of \eqn{X_j}, see \code{\link{Check.manifold}}.}
      \item{p}{the number of \eqn{X_j}.}
}}
}
\value{
an \eqn{n'\times m} matrix of predicted values \eqn{\hat{Y}_{new}}.
}
\description{
Predict \eqn{\hat{Y}_{new}} for the given \eqn{X_{new}} using an \code{\link{LM}} object.
}
