% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_list.R
\name{predict.PCA.manifold.list}
\alias{predict.PCA.manifold.list}
\title{Prediction of Score Matrices for a List of Manifold-valued Data}
\usage{
\method{predict}{PCA.manifold.list}(object, Xdatanew)
}
\arguments{
\item{object}{a \code{\link{PCA.manifold.list}} object.}

\item{Xdatanew}{a new list of covariates with the following arguments (see also \code{\link{covariates.generate}}):
\describe{
      \item{j}{a \eqn{p} list of manifold-valued covariates, where each \eqn{j}th element is an \eqn{n\times T_j} matrix.}
      \item{spaces}{a \eqn{p} vector of the underlying spaces \eqn{\mathcal{M}_j} of \eqn{X_j}, see \code{\link{Check.manifold}}.}
      \item{p}{the number of \eqn{X_j}.}
}}
}
\value{
a \eqn{p} list of score matrices.
}
\description{
Computes a list of score matrices for a list of the manifold-valued data based on \code{\link{PCA.manifold.list}}.
}
