% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_list.R
\name{predict.PCA.manifold.list}
\alias{predict.PCA.manifold.list}
\title{Prediction of principal component score for manifold-valued data}
\usage{
\method{predict}{PCA.manifold.list}(object, Xdatanew)
}
\arguments{
\item{object}{a object obtained from \code{\link{PCA.manifold.list}}.}

\item{Xdatanew}{a \eqn{p} list of new manifold-valued data.}
}
\value{
a \eqn{p} list of score matrices, see \code{\link{predict.PCA.manifold}}.
}
\description{
Prediction of principal component scores for a list of manifold-valued data.
}
