% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_Wasserstein.R, R/Geometry_functional.R,
%   R/Geometry_simplex.R, R/Geometry_sphere.R
\name{predict.PCA.manifold}
\alias{predict.PCA.manifold}
\alias{predict.PCA.BayesHilbert}
\alias{predict.PCA.Euclid}
\alias{predict.PCA.Wasserstein}
\alias{predict.PCA.functional}
\alias{predict.PCA.simplex}
\alias{predict.PCA.sphere}
\title{Prediction score matrix of manifold-valued data.}
\usage{
\method{predict}{PCA.manifold}(object, Xnew)

\method{predict}{PCA.BayesHilbert}(object, Xnew)

\method{predict}{PCA.Euclid}(object, Xnew)

\method{predict}{PCA.Wasserstein}(object, Xnew)

\method{predict}{PCA.functional}(object, Xnew)

\method{predict}{PCA.simplex}(object, Xnew)

\method{predict}{PCA.sphere}(object, Xnew)
}
\arguments{
\item{object}{a PCA.manifold object created by \code{\link{PCA.manifold}}}

\item{Xnew}{a matrix with dimension (n,m). Each row is a point on the space.}
}
\value{
a score matrix with dimension (n,m')
}
\description{
Prediction score matrix of manifold-valued data.

Prediction score matrix for Bayes-Hilbert data

Prediction score matrix for Euclid data

Prediction score matrix for Wasserstein data

Prediction score matrix for functional data

Prediction score matrix for simplex data

Prediction score matrix for sphere data
}
\section{Functions}{
\itemize{
\item \code{predict(PCA.BayesHilbert)}: Method

\item \code{predict(PCA.Euclid)}: Method

\item \code{predict(PCA.Wasserstein)}: Method

\item \code{predict(PCA.functional)}: Method

\item \code{predict(PCA.simplex)}: Method

\item \code{predict(PCA.sphere)}: Method

}}
