% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_AffInv.R, R/Geometry_SPD_LogEuclid.R,
%   R/Geometry_Wasserstein.R, R/Geometry_functional.R, R/Geometry_simplex.R,
%   R/Geometry_sphere.R
\name{predict.PCA.manifold}
\alias{predict.PCA.manifold}
\alias{predict.PCA.BayesHilbert}
\alias{predict.PCA.Euclid}
\alias{predict.PCA.SPD.AffInv}
\alias{predict.PCA.SPD.LogEuclid}
\alias{predict.PCA.Wasserstein}
\alias{predict.PCA.functional}
\alias{predict.PCA.simplex}
\alias{predict.PCA.sphere}
\title{Prediction score matrix of manifold-valued data.}
\usage{
\method{predict}{PCA.manifold}(object, Xnew)

\method{predict}{PCA.BayesHilbert}(object, Xnew)

\method{predict}{PCA.Euclid}(object, Xnew)

\method{predict}{PCA.SPD.AffInv}(object, Xnew)

\method{predict}{PCA.SPD.LogEuclid}(object, Xnew)

\method{predict}{PCA.Wasserstein}(object, Xnew)

\method{predict}{PCA.functional}(object, Xnew)

\method{predict}{PCA.simplex}(object, Xnew)

\method{predict}{PCA.sphere}(object, Xnew)
}
\arguments{
\item{object}{a \code{\link{PCA.manifold}} object.}

\item{Xnew}{an \eqn{n\times m} matrix). Each row is a point on the space.}
}
\value{
an \eqn{n\times m'} score matrix.
}
\description{
Prediction score matrix of manifold-valued data.

Prediction score matrix for Bayes-Hilbert data

Prediction score matrix for Euclid data

Prediction score matrix for SPD data with the Affine-invariant metric

Prediction score matrix for SPD data with the Log-Euclidean metric

Prediction score matrix for Wasserstein data

Prediction score matrix for functional data

Prediction score matrix for simplex data

Prediction score matrix for sphere data
}
\section{Functions}{
\itemize{
\item \code{predict(PCA.BayesHilbert)}: Method

\item \code{predict(PCA.Euclid)}: Method

\item \code{predict(PCA.SPD.AffInv)}: Method

\item \code{predict(PCA.SPD.LogEuclid)}: Method

\item \code{predict(PCA.Wasserstein)}: Method

\item \code{predict(PCA.functional)}: Method

\item \code{predict(PCA.simplex)}: Method

\item \code{predict(PCA.sphere)}: Method

}}
