% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM.R
\name{reduce.dimension}
\alias{reduce.dimension}
\title{Demension reduction function for a list of X.}
\usage{
reduce.dimension(X, Xdim.max = 200, margin = 1)
}
\arguments{
\item{X}{a \eqn{p} list of matrices. Each \eqn{X_j} is a \eqn{n\times d_j} matrix.}

\item{Xdim.max}{a max dimension of \eqn{X_j}, int>0.}

\item{margin}{the direction which the dimension reduction applied over.}
}
\value{
a \eqn{p} list of \eqn{X_j'}.
}
\description{
Change the dimension of \eqn{X_j} at most \eqn{min(d_j,Xdim.max)}.
}
