% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_AffInv.R, R/Geometry_SPD_LogEuclid.R,
%   R/Geometry_Wasserstein.R, R/Geometry_functional.R, R/Geometry_simplex.R,
%   R/Geometry_sphere.R
\name{RieExp.manifold}
\alias{RieExp.manifold}
\alias{RieExp.BayesHilbert}
\alias{RieExp.Euclid}
\alias{RieExp.SPD.AffInv}
\alias{RieExp.SPD.LogEuclid}
\alias{RieExp.Wasserstein}
\alias{RieExp.functional}
\alias{RieExp.simplex}
\alias{RieExp.sphere}
\title{Riemannian exponential map on the manifold}
\usage{
RieExp.manifold(p, u, space = "Euclid")

RieExp.BayesHilbert(p, u)

RieExp.Euclid(p, u)

RieExp.SPD.AffInv(p, u)

RieExp.SPD.LogEuclid(p, u)

RieExp.Wasserstein(p, u)

RieExp.functional(p, u)

RieExp.simplex(p, u)

RieExp.sphere(p, u)
}
\arguments{
\item{p}{an \eqn{m} vector representing the base point in \eqn{\mathcal{M}}.}

\item{u}{an \eqn{n\times m'} matrix, where each row is a tangent vector on \eqn{T_p\mathcal{M}}.}

\item{space}{the name of space, which must be one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n\times m} matrix of Riemannian exponentials of each row of \eqn{u}.
}
\description{
Compute the Riemnnain exponential map on the manifold \eqn{\mathcal{M}}.
}
\section{Functions}{
\itemize{
\item \code{RieExp.BayesHilbert()}: Method

\item \code{RieExp.Euclid()}: Method

\item \code{RieExp.SPD.AffInv()}: Method

\item \code{RieExp.SPD.LogEuclid()}: Method

\item \code{RieExp.Wasserstein()}: Method

\item \code{RieExp.functional()}: Method

\item \code{RieExp.simplex()}: Method

\item \code{RieExp.sphere()}: Method

}}
