% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_AffInv.R, R/Geometry_SPD_LogEuclid.R,
%   R/Geometry_Wasserstein.R, R/Geometry_functional.R, R/Geometry_simplex.R,
%   R/Geometry_sphere.R
\name{RieExp.manifold}
\alias{RieExp.manifold}
\alias{RieExp.BayesHilbert}
\alias{RieLog.BayesHilbert}
\alias{RieExp.Euclid}
\alias{RieLog.Euclid}
\alias{RieExp.SPD.AffInv}
\alias{RieLog.SPD.AffInv}
\alias{RieExp.SPD.LogEuclid}
\alias{RieLog.SPD.LogEuclid}
\alias{RieExp.Wasserstein}
\alias{RieLog.Wasserstein}
\alias{RieExp.functional}
\alias{RieLog.functional}
\alias{RieExp.simplex}
\alias{RieLog.simplex}
\alias{RieExp.sphere}
\alias{RieLog.sphere}
\title{Riemannian exponential map}
\usage{
RieExp.manifold(p, u, space = "Euclid")

RieExp.BayesHilbert(p, u)

RieLog.BayesHilbert(p, q)

RieExp.Euclid(p, u)

RieLog.Euclid(p, q)

RieExp.SPD.AffInv(p, u)

RieLog.SPD.AffInv(p, q)

RieExp.SPD.LogEuclid(p, u)

RieLog.SPD.LogEuclid(p, q)

RieExp.Wasserstein(p, u)

RieLog.Wasserstein(p, q)

RieExp.functional(p, u)

RieLog.functional(p, q)

RieExp.simplex(p, u)

RieLog.simplex(p, q)

RieExp.sphere(p, u)

RieLog.sphere(p, q)
}
\arguments{
\item{p}{an \eqn{m} vector of the base point on tangent space}

\item{u}{an \eqn{n\times m'} matrix. Each row is a tangent vector.}

\item{space}{a name of space, one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n\times m} matrix. Each row is a point on the Riemannian metric spaces.
}
\description{
Riemannian exponential map

Riemannian exponential map for Bayes-Hilbert data

Riemannian logarithmic map for Bayes-Hilbert data

Riemannian exponential map for Euclid data

Riemannian logarithmic map for Euclid data

Riemannian exponential map for SPD data with the Affine-invariant metric

Riemannian logarithmic map for SPD data with the Affine-invariant metric

Riemannian exponential map for SPD data with the Log-Euclidean metric

Riemannian logarithmic map for SPD data with the Log-Euclidean metric

Riemannian exponential map for Wasserstein data

Riemannian logarithmic map for Wasserstein data

Riemannian exponential map for functional data

Riemannian logarithmic map for functional data

Riemannian exponential map for simplex data

Riemannian logarithmic map for simplex data

Riemannian exponential map for sphere data

Riemannian logarithmic map for sphere data
}
\section{Functions}{
\itemize{
\item \code{RieExp.BayesHilbert()}: Method

\item \code{RieLog.BayesHilbert()}: Method

\item \code{RieExp.Euclid()}: Method

\item \code{RieLog.Euclid()}: Method

\item \code{RieExp.SPD.AffInv()}: Method

\item \code{RieLog.SPD.AffInv()}: Method

\item \code{RieExp.SPD.LogEuclid()}: Method

\item \code{RieLog.SPD.LogEuclid()}: Method

\item \code{RieExp.Wasserstein()}: Method

\item \code{RieLog.Wasserstein()}: Method

\item \code{RieExp.functional()}: Method

\item \code{RieLog.functional()}: Method

\item \code{RieExp.simplex()}: Method

\item \code{RieLog.simplex()}: Method

\item \code{RieExp.sphere()}: Method

\item \code{RieLog.sphere()}: Method

}}
