% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R, R/Geometry_BayesHilbert.R,
%   R/Geometry_Euclid.R, R/Geometry_SPD_AffInv.R, R/Geometry_SPD_LogEuclid.R,
%   R/Geometry_Wasserstein.R, R/Geometry_functional.R, R/Geometry_simplex.R,
%   R/Geometry_sphere.R
\name{RieLog.manifold}
\alias{RieLog.manifold}
\alias{RieLog.BayesHilbert}
\alias{RieLog.Euclid}
\alias{RieLog.SPD.AffInv}
\alias{RieLog.SPD.LogEuclid}
\alias{RieLog.Wasserstein}
\alias{RieLog.functional}
\alias{RieLog.simplex}
\alias{RieLog.sphere}
\title{Riemannian logarithmic map on the manifold}
\usage{
RieLog.manifold(p, q, space = "Euclid")

RieLog.BayesHilbert(p, q)

RieLog.Euclid(p, q)

RieLog.SPD.AffInv(p, q)

RieLog.SPD.LogEuclid(p, q)

RieLog.Wasserstein(p, q)

RieLog.functional(p, q)

RieLog.simplex(p, q)

RieLog.sphere(p, q)
}
\arguments{
\item{p, q}{\eqn{m} vectors or \eqn{n\times m} matrices, where each row is a point on \eqn{\mathcal{M}}.}

\item{space}{the name of space, which must be one of 'Euclid', 'simplex', 'sphere', 'SPD.LogEuclid', 'SPD.AffInv', 'functional', 'BayesHilbert', or 'Wasserstein'.}
}
\value{
an \eqn{n\times m} matrix of Riemannian logarithmics from  corresponding rows of \eqn{p} to those of \eqn{q}.
}
\description{
Compute the Riemnnain logarithmic map on the manifold \eqn{\mathcal{M}}.
}
\section{Functions}{
\itemize{
\item \code{RieLog.BayesHilbert()}: Method

\item \code{RieLog.Euclid()}: Method

\item \code{RieLog.SPD.AffInv()}: Method

\item \code{RieLog.SPD.LogEuclid()}: Method

\item \code{RieLog.Wasserstein()}: Method

\item \code{RieLog.functional()}: Method

\item \code{RieLog.simplex()}: Method

\item \code{RieLog.sphere()}: Method

}}
