% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_data_generation.R
\name{tensor.beta.generate.each}
\alias{tensor.beta.generate.each}
\title{Generate linear operators from \eqn{H_1} to \eqn{H_2}}
\usage{
tensor.beta.generate.each(Xspace, Yspace, Xdim, Ydim, beta.norm = 1)
}
\arguments{
\item{Xspace}{an underlying space of \eqn{H_1}.}

\item{Yspace}{an underlying space of \eqn{H_2}.}

\item{Xdim}{an intrinsic dimension of \eqn{H_1}.}

\item{Ydim}{an intrinsic dimension of \eqn{H_2}.}

\item{beta.norm}{the tensor norm.}
}
\value{
a tensor product element in \eqn{H_1\otimes H_2} generated by \code{\link{make.tensor}}.
}
\description{
Generate a linear operator \eqn{B_j} from \eqn{H_1} to \eqn{H_2}. 
It can be identified as an element in the tensor product space \eqn{H_1\otimes H_2}.
}
