% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_data_generation.R
\name{tensor.beta.generate}
\alias{tensor.beta.generate}
\title{Generate a list of operators from \eqn{H_j} to \eqn{H_Y}.}
\usage{
tensor.beta.generate(
  Xspaces,
  Yspace,
  Xdims,
  Ydim,
  proper.indices = NULL,
  beta.norm = 1,
  ...
)
}
\arguments{
\item{Xspaces}{a \eqn{p} vector of underlying spaces of \eqn{X_j}.}

\item{Yspace}{an underlying space of \eqn{Y}.}

\item{Xdims}{a \eqn{p} vector of intrinsic dimensions of \eqn{H_j}.}

\item{Ydim}{an intrinsic dimension of \eqn{H_Y}.}

\item{proper.indices}{an indices of nonzero operators.}

\item{beta.norm}{the tensor norm.}
}
\value{
a list of operators,
   \describe{
      \item{j}{a \eqn{p} list of generated data. Each jth element is an \eqn{n}-by-\eqn{Xdim_j} matrix.}
      \item{Xspaces}{a \eqn{p} vector of spaces of \eqn{X_j}.}
      \item{p}{a number of \eqn{X_j}.}
}
}
\description{
Generate a list of operators \eqn{B_j} from \eqn{H_j} to \eqn{H_Y}.
Each operator can be identified as an element in the tensor product space \eqn{H_j\otimes H_Y}.
}
