% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vector.norm}
\alias{vector.norm}
\title{vector norm of matrices on the tangent space at \eqn{p}}
\usage{
vector.norm(X, p, space = "Euclid", type = "L2")
}
\arguments{
\item{X}{a \eqn{p}-by-\eqn{m} matrix of manifold-valued data.}

\item{p}{a base point on the tangent space.}

\item{space}{an underlying space of \eqn{X}.}

\item{type}{a type of vector norm. One of 'L2', 'L1', or 'Linf'.}
}
\value{
a vector norm
}
\description{
vector norm of matrices on the tangent space at \eqn{p}
}
