% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation.R
\name{Xmu.generate}
\alias{Xmu.generate}
\title{Generate Frechet means of \eqn{X_j}}
\usage{
Xmu.generate(m, space)
}
\arguments{
\item{m}{the length of each observation. Typically, this equals the dimension, but it can differ in some cases.}

\item{space}{the name of space, see \code{\link{Check.manifold}}.}
}
\value{
the Frechet mean \eqn{\mu_j} of \eqn{X_j}, an \eqn{m} vector.
}
\description{
Generate Frechet means \eqn{\mu_j} of \eqn{X_j}.
}
