% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_data_generation.R
\name{Xmu.generate}
\alias{Xmu.generate}
\title{Generate Frechet means of \eqn{X_j}}
\usage{
Xmu.generate(m, space)
}
\arguments{
\item{m}{length of each observation.}

\item{space}{a name of space, see \code{\link{Check.manifold}}.}
}
\value{
a \eqn{m} vector.
}
\description{
Generate Frechet means of \eqn{X_j}
}
