% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_generation.R
\name{Ymu.generate}
\alias{Ymu.generate}
\title{Generate the Frechet mean for \eqn{Y}}
\usage{
Ymu.generate(m, space, ngrid = 100)
}
\arguments{
\item{m}{the length of each observation. Typically, this equals the dimension, but it can differ in some cases.}

\item{space}{the name of space, see \code{\link{Check.manifold}}.}

\item{ngrid}{the number of grids. This is only used for infinite-dimensional spaces, see \code{\link{Check.manifold}}: "functional", "BayesHilbert" and "Wasserstein" spaces.}
}
\value{
The Frechet mean \eqn{\mu_Y} of \eqn{Y}, an \eqn{m} vector.
}
\description{
This code is the same as \code{\link{Xmu.generate}}.
}
