% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_data_generation.R
\name{Ymu.generate}
\alias{Ymu.generate}
\title{Generate the Frechet means for \eqn{Y}}
\usage{
Ymu.generate(m, space)
}
\arguments{
\item{m}{length of each observation.}

\item{space}{a name of space, see \code{\link{Check.manifold}}.}
}
\value{
an \eqn{m} vector.
}
\description{
The code is the same as \code{\link{Xmu.generate}}.
}
